% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_annotation.R
\name{ann_arrow}
\alias{ann_arrow}
\title{Add an "arrow" annotation to a Bokeh figure}
\usage{
ann_arrow(fig, x_start = NULL, y_start = NULL, x_end = NULL,
  y_end = NULL, color = NULL, alpha = NULL, width = NULL,
  data = figure_data(fig), start = NULL, start_units = NULL, end = NULL,
  end_units = NULL, line_dash = NULL, line_join = NULL,
  line_dash_offset = NULL, line_cap = NULL, lgroup = NULL, lname = NULL)
}
\arguments{
\item{fig}{Figure to modify.}

\item{x_start}{The x-coordinates to locate the start of the arrows.}

\item{y_start}{The y-coordinates to locate the start of the arrows.}

\item{x_end}{The x-coordinates to locate the end of the arrows.}

\item{y_end}{The y-coordinates to locate the end of the arrows.}

\item{width}{The line width values for the arrow body.}

\item{data}{An optional data frame supplying data to which other parameters can refer.}

\item{start}{Instance of ArrowHead.}

\item{start_units}{The unit type for the start_x and start_y attributes. Interpreted as "data space" units by default.}

\item{end}{Instance of ArrowHead.}

\item{end_units}{The unit type for the end_x and end_y attributes. Interpreted as "data space" units by default.}

\item{line_dash}{The line dash values for the arrow body.}

\item{line_join}{The line join values for the arrow body.}

\item{line_dash_offset}{The line dash offset values for the arrow body.}

\item{line_cap}{The line cap values for the arrow body.}
}
\description{
Add an "arrow" annotation to a Bokeh figure
}
\examples{
figure() \%>\%
  ly_points(1:10, 1:10) \%>\%
  ann_arrow(3, 5, 4, 4)

# with different head
figure() \%>\%
  ly_points(1:10, 1:10) \%>\%
  ann_arrow(3, 5, 4, 4, end = "vee")

# with more specific head parameters using "arrow()"
figure() \%>\%
  ly_points(1:10, 1:10) \%>\%
  ann_arrow(3, 5, 4, 4, end = arrow("vee", color = "red"))

# vectorized example using a data frame of values as input
da <- cbind(expand.grid(1:4, 3:8), cbind(expand.grid(2:5, 4:9)))
names(da) <- c("x1", "y1", "x2", "y2")
figure() \%>\%
  ly_points(x = 1:5, y = 3:7) \%>\%
  ann_arrow(x1, y1, x2, y2, data = da,
    end = arrow("vee", color = "red", size = 50), start = arrow("tee", color = "red"))
}
\seealso{
Other annotation functions: \code{\link{ann_band}},
  \code{\link{ann_box}}, \code{\link{ann_labels}},
  \code{\link{ann_poly}}, \code{\link{ann_span}},
  \code{\link{ann_title}}, \code{\link{ann_whisker}},
  \code{\link{arrow}}
}
