% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_annotation.R
\name{ann_labels}
\alias{ann_labels}
\title{Add a "label" annotation to a Bokeh figure}
\usage{
ann_labels(fig, x = NULL, y = NULL, text = NULL, angle = NULL,
  data = figure_data(fig), x_units = NULL, y_units = NULL,
  angle_units = NULL, x_offset = NULL, y_offset = NULL,
  text_align = NULL, text_color = NULL, text_alpha = NULL,
  text_font = NULL, text_font_size = NULL, text_font_style = NULL,
  text_baseline = NULL, background_fill_color = NULL,
  background_fill_alpha = NULL, border_line_color = NULL,
  border_line_alpha = NULL, border_line_width = NULL,
  border_line_join = NULL, border_line_dash = NULL,
  border_line_dash_offset = NULL, border_line_cap = NULL,
  render_mode = NULL, lgroup = NULL, lname = NULL)
}
\arguments{
\item{fig}{Figure to modify.}

\item{x}{The x-coordinates to locate the text anchors.}

\item{y}{The y-coordinates to locate the text anchors.}

\item{text}{The text values to render.}

\item{angle}{The angles to rotate the text, as measured from the horizontal. Warning: The center of rotation for canvas and css render_modes is different. For \code{render_mode="canvas"} the label is rotated from the top-left corner of the annotation, while for \code{render_mode="css"} the annotation is rotated around it's center.}

\item{data}{An optional data frame supplying data to which other parameters can refer.}

\item{x_units}{The unit type for the xs attribute. Interpreted as "data space" units by default. One of "data" or "screen".}

\item{y_units}{The unit type for the ys attribute. Interpreted as "data space" units by default. One of "data" or "screen".}

\item{angle_units}{One of "rad" or "deg".}

\item{x_offset}{Offset values to apply to the x-coordinates. This is useful, for instance, if it is desired to "float" text a fixed distance in screen units from a given data position.}

\item{y_offset}{Offset values to apply to the y-coordinates. This is useful, for instance, if it is desired to "float" text a fixed distance in screen units from a given data position.}

\item{text_align}{The text align values for the text.}

\item{text_color}{The text color values for the text.}

\item{text_alpha}{The text alpha values for the text.}

\item{text_font}{The text font values for the text.}

\item{text_font_size}{The text font size values for the text.}

\item{text_font_style}{The text font style values for the text.}

\item{text_baseline}{The text baseline values for the text.}

\item{background_fill_color}{The fill color values for the text bounding box.}

\item{background_fill_alpha}{The fill alpha values for the text bounding box.}

\item{border_line_color}{The line color values for the text bounding box.}

\item{border_line_alpha}{The line alpha values for the text bounding box.}

\item{border_line_width}{The line width values for the text bounding box.}

\item{border_line_join}{The line join values for the text bounding box.}

\item{border_line_dash}{The line dash values for the text bounding box.}

\item{border_line_dash_offset}{The line dash offset values for the text bounding box.}

\item{border_line_cap}{The line cap values for the text bounding box.}

\item{render_mode}{Specifies whether the text is rendered as a canvas element or as an css element overlaid on the canvas. The default mode is "canvas". Note: The CSS labels won't be present in the output using the "save" tool. Warning: Not all visual styling properties are supported if the render_mode is set to "css". The border_line_dash property isn't fully supported and border_line_dash_offset isn't supported at all. Setting text_alpha will modify the opacity of the entire background box and border in addition to the text. Finally, clipping label annotations inside of the plot area isn't supported in "css" mode.}

\item{lgroup}{Layer group.}

\item{lname}{Layer name.}
}
\description{
Add a "label" annotation to a Bokeh figure
}
\examples{
d <- data.frame(x = 1:10, y = rnorm(10), txt = letters[1:10])
figure(data = d) \%>\%
  ly_points(x, y) \%>\%
  ann_labels(x, y, txt)
}
\seealso{
Other annotation functions: \code{\link{ann_arrow}},
  \code{\link{ann_band}}, \code{\link{ann_box}},
  \code{\link{ann_poly}}, \code{\link{ann_span}},
  \code{\link{ann_title}}, \code{\link{ann_whisker}},
  \code{\link{arrow}}
}
