% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_annotation.R
\name{ann_span}
\alias{ann_span}
\title{Add a "span" (horizontal or vertical line) annotation to a Bokeh figure}
\usage{
ann_span(fig, location = NULL, dimension = c("height", "width"),
  location_units = NULL, line_color = NULL, line_alpha = NULL,
  line_width = NULL, line_cap = NULL, line_join = NULL,
  line_dash = NULL, line_dash_offset = NULL, render_mode = NULL,
  lgroup = NULL, lname = NULL)
}
\arguments{
\item{fig}{Figure to modify.}

\item{location}{The location of the span, along "dimension".}

\item{dimension}{The direction of the span.}

\item{location_units}{The unit type for the location attribute. Interpreted as "data space" units by default.}

\item{line_color}{The line color values for the span.}

\item{line_alpha}{The line alpha values for the span.}

\item{line_width}{The line width values for the span.}

\item{line_cap}{The line cap values for the span.}

\item{line_join}{The line join values for the span.}

\item{line_dash}{The line dash values for the span.}

\item{line_dash_offset}{The line dash offset values for the span.}

\item{render_mode}{Specifies whether the span is rendered as a canvas element or as a css element overlaid on the canvas. The default mode is "canvas". Warning: The line_dash and line_dash_offset attributes aren't supported if the render_mode is set to "css".}

\item{lgroup}{Layer group.}

\item{lname}{Layer name.}
}
\description{
Add a "span" (horizontal or vertical line) annotation to a Bokeh figure
}
\seealso{
Other annotation functions: \code{\link{ann_arrow}},
  \code{\link{ann_band}}, \code{\link{ann_box}},
  \code{\link{ann_labels}}, \code{\link{ann_poly}},
  \code{\link{ann_title}}, \code{\link{ann_whisker}},
  \code{\link{arrow}}
}
