% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_annotation.R
\name{ann_whisker}
\alias{ann_whisker}
\title{Add a "whisker" annotation to a Bokeh figure}
\usage{
ann_whisker(fig, base = NULL, lower = NULL, upper = NULL,
  lower_head = NULL, upper_head = NULL, dimension = NULL,
  data = figure_data(fig), base_units = NULL, lower_units = NULL,
  upper_units = NULL, line_color = NULL, line_alpha = NULL,
  line_width = NULL, line_dash = NULL, line_dash_offset = NULL,
  line_join = NULL, line_cap = NULL, lgroup = NULL, lname = NULL)
}
\arguments{
\item{fig}{Figure to modify.}

\item{base}{The orthogonal coordinates of the upper and lower values.}

\item{lower}{The coordinates of the lower end of the whiskers.}

\item{upper}{The coordinations of the upper end of the whiskers.}

\item{lower_head}{Instance of ArrowHead.}

\item{upper_head}{Instance of ArrowHead.}

\item{dimension}{The direction of the band.}

\item{data}{An optional data frame supplying data to which other parameters can refer.}

\item{base_units}{Either "screen" or "data".}

\item{lower_units}{Either "screen" or "data".}

\item{upper_units}{Either "screen" or "data".}

\item{line_color}{The line color values for the whisker body.}

\item{line_alpha}{The line alpha values for the whisker body.}

\item{line_width}{The line width values for the whisker body.}

\item{line_dash}{The line dash values for the whisker body.}

\item{line_dash_offset}{The line dash offset values for the whisker body.}

\item{line_join}{The line join values for the whisker body.}

\item{line_cap}{The line cap values for the whisker body.}

\item{lgroup}{Layer group.}

\item{lname}{Layer name.}
}
\description{
Add a "whisker" annotation to a Bokeh figure
}
\seealso{
Other annotation functions: \code{\link{ann_arrow}},
  \code{\link{ann_band}}, \code{\link{ann_box}},
  \code{\link{ann_labels}}, \code{\link{ann_poly}},
  \code{\link{ann_span}}, \code{\link{ann_title}},
  \code{\link{arrow}}
}
