% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{asis}
\alias{asis}
\title{Specify that a plot variable should be treated "as-is" instead of being mapped.}
\usage{
asis(x)
}
\arguments{
\item{x}{A vector or variable name.}
}
\description{
Specify that a plot variable should be treated "as-is" instead of being mapped.
}
