% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer__utils.R
\name{b_eval}
\alias{b_eval}
\title{Eval lazy symbol}
\usage{
b_eval(data)
}
\arguments{
\item{data}{data set to be used for evaluation.  May be \code{NULL}}
}
\value{
a function that takes in one lazy argument to be evaluated
}
\description{
Evaluate the argument from the env it came from, or from within the data.  The arg supplied to the returned function must be lazy.
}

