% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer__utils.R
\name{b_xy_data_and_names2}
\alias{b_xy_data_and_names2}
\title{process data and data names}
\usage{
b_xy_data_and_names2(d2, xlab, ylab)
}
\arguments{
\item{xlab}{xlab argument}

\item{ylab}{ylab argument}

\item{dt}{only the first two components of the data section}
}
\value{
a list containing d2 and xyNames. d2 is the first two data positions and xyNames is a list of xName and yName
}
\description{
process data and data names
}

