% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_helpers.R
\name{bokeh_render_json}
\alias{bokeh_render_json}
\title{Plot a Bokeh JSON specification}
\usage{
bokeh_render_json(json, is_file = TRUE)
}
\arguments{
\item{json}{path to json file}

\item{is_file}{is \code{json} a json string (FALSE) or a path to a json file (TRUE)?}
}
\description{
Take a path to a Bokeh JSON plot specification file and render it in the browser.
}
\note{
This is mainly useful for development / debugging purposes for reading in json created from another platform like Python, or to be used with tweaking json output from \code{\link{print_model_json}}.
}
\seealso{
\code{\link{print_model_json}}
}
