% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_callback.R
\name{console_callback}
\alias{console_callback}
\title{Specify a console callback}
\usage{
console_callback()
}
\description{
This registers a callback that simply prints the callback objects in the javascript console of your web browser.  A probalby more useful callback is the \code{\link{debug_callback}} which will place you inside a debugger in your web browser allowing you to inspect the callback objects.
}
\examples{
\donttest{
figure() \%>\%
  ly_points(1:10) \%>\%
  x_range(callback = console_callback()) \%>\%
  y_range(callback = console_callback())
}
}

