% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{custom_js_transform}
\alias{custom_js_transform}
\title{Specify a custom JavaScript transformation.}
\usage{
custom_js_transform(func, v_func = NULL, global = NULL)
}
\arguments{
\item{func}{A string of JavaScript code to be applied to one value of an array.}

\item{v_func}{An optional string of JavaScript code to be applied to an entire array. If not specified, the array will be looped over and \code{func} will be applied to each element.}

\item{global}{A string of JavaScript code indicating global values to be specified prior to the code for \code{func} or \code{v_func}.}
}
\description{
Specify a custom JavaScript transformation.
}
