% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer__utils.R
\name{data_name_list}
\alias{data_name_list}
\title{List of all types of data name structures that could appear}
\usage{
data_name_list()
}
\description{
List of all types of data name structures that could appear
}
