% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_callback.R
\name{debug_callback}
\alias{debug_callback}
\title{Specify a "debug" callback}
\usage{
debug_callback(lnames = NULL, args = NULL)
}
\arguments{
\item{lnames}{vector of layer names to be made available inside the callback in addition to the default callback objects (see \code{\link{custom_callback}} for details)}

\item{args}{named list of additional references to objects to be addressable in the callback}
}
\description{
This registers a callback that simply places you inside a debugger in your web browser allowing you to inspect the callback objects.
}
\examples{
\donttest{
figure() \%>\%
  ly_points(1:10, lname = "points") \%>\%
  tool_tap(debug_callback("points"), "points")
}
}

