% Generated by roxygen2 (4.0.2): do not edit by hand
\name{figure}
\alias{figure}
\title{Start a Bokeh Figure}
\usage{
figure(width = 480, height = 480, title = NULL, xlab = NULL,
  ylab = NULL, xlim = NULL, ylim = NULL, padding_factor = 0.07,
  plot_width = NULL, plot_height = NULL, xgrid = TRUE, ygrid = TRUE,
  xaxes = "below", yaxes = "left", tools = c("pan", "wheel_zoom",
  "box_zoom", "resize", "reset", "save"), theme = getOption("bokeh_theme"))
}
\arguments{
\item{width}{figure width in pixels}

\item{height}{figure width in pixels}

\item{title}{a title to display above the plot. - "title" is also the prefix for a set of Text Properties, so you can set the font for the title with the parameter text_font.}

\item{xlab}{label for x axis}

\item{ylab}{label for y axis}

\item{xlim}{the extent of the plotting area in the x-dimension (will be computed automatically if not specified).}

\item{ylim}{the extent of the plotting area in the y-dimension (will be computed automatically if not specified).}

\item{padding_factor}{if limits are not specified, by what factor should the extents of the data be padded}

\item{plot_width,plot_height}{width and height of the entire plot in pixels, including border space}

\item{xgrid}{whether to draw x axis grid lines}

\item{ygrid}{whether to draw y axis grid lines}

\item{xaxes}{where to put x axis labels}

\item{yaxes}{where to put y axis labels}

\item{tools}{interactivity tools options}

\item{theme}{an rbokeh theme to use (tableau by default)}
}
\description{
Start a Bokeh Figure
}

