% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbokeh-package.R
\docType{data}
\name{flightfreq}
\alias{flightfreq}
\title{Flight frequency dataset}
\usage{
flightfreq
}
\description{
Daily counts of domestic flights in the U.S. from 1999 to mid-2008
}
\examples{
\donttest{
p <- figure(width = 1000) \%>\%
  ly_points(date, Freq, data = flightfreq,
    hover = list(date, Freq, dow), size = 5) \%>\%
  ly_abline(v = as.Date("2001-09-11"))
p
}
}
\keyword{data}
