% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_ajax_data.R
\name{get_ajax_sample}
\alias{get_ajax_sample}
\title{Get a Sample of Data from Ajax Data Source}
\usage{
get_ajax_sample(x)
}
\arguments{
\item{x}{object obtained from \code{\link{ajax_data}}.}
}
\description{
Get a Sample of Data from Ajax Data Source
}
\examples{
d <- ajax_data("http://127.0.0.1:8080")
\dontrun{
get_ajax_sample(d)
}
}
