% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/json_helpers.R
\name{get_bokeh_html}
\alias{get_bokeh_html}
\title{Get the HTML content required to embed a Bokeh figure}
\usage{
get_bokeh_html(fig)
}
\arguments{
\item{fig}{figure}
}
\description{
Get the HTML content required to embed a Bokeh figure
}

