% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_object_refs}
\alias{get_object_refs}
\title{Get object ids and types from a figure}
\usage{
get_object_refs(fig)
}
\arguments{
\item{fig}{a figure object}
}
\description{
Get object ids and types from a figure
}
