% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_gmap.R
\name{gmap}
\alias{gmap}
\title{Initialize a Bokeh Google Map plot}
\usage{
gmap(data = NULL, lat = 0, lng = 0, zoom = 0, api_key = NULL,
  map_type = "hybrid", map_style = NULL, scale_control = NULL,
  width = 600, height = 600, title = NULL, title_location = NULL,
  xlab = NULL, ylab = NULL, xlim = NULL, ylim = NULL,
  range_padding = 0.07, xgrid = FALSE, ygrid = FALSE,
  legend_location = "top_right", logo = NULL,
  theme = getOption("bokeh_theme"), tools = c("pan", "wheel_zoom", "save"),
  toolbar_location = "above", toolbar_sticky = NULL, h_symmetry = TRUE,
  v_symmetry = FALSE, lod_factor = 10, lod_interval = 300,
  lod_threshold = NULL, lod_timeout = 500, output_backend = "canvas",
  min_border = NULL, min_border_left = NULL, min_border_bottom = NULL,
  min_border_right = NULL, min_border_top = NULL, hidpi = NULL)
}
\arguments{
\item{data}{A default data frame to be used by the layers added to the figure.}

\item{lat}{The latitude where the map should be centered.}

\item{lng}{The longitude where the map should be centered.}

\item{zoom}{The initial zoom level to use when displaying the map.}

\item{api_key}{Google Maps \href{https://developers.google.com/maps/documentation/javascript/get-api-key}{API key}. If not specified, a search for an environmen variable, \code{GMAP_API_KEY}, will be done.}

\item{map_type}{The `map type`_ to use for the GMapPlot. One of 'satellite', 'roadmap', 'terrain', 'hybrid'. See \url{https://developers.google.com/maps/documentation/javascript/reference#MapTypeId} for more details.}

\item{map_style}{a json string of a Google Maps style - see \code{\link{gmap_style}}}

\item{scale_control}{Whether the Google map should display its distance scale control.}

\item{width}{TODO_DOC}

\item{height}{TODO_DOC}

\item{title}{TODO_DOC}

\item{title_location}{Where the title will be located. Titles on the left or right side will  be rotated. One of 'above', 'below', 'left', 'right'.}

\item{xlab}{TODO_DOC}

\item{ylab}{TODO_DOC}

\item{xlim}{TODO_DOC}

\item{ylim}{TODO_DOC}

\item{range_padding}{TODO_DOC}

\item{xgrid}{TODO_DOC}

\item{ygrid}{TODO_DOC}

\item{legend_location}{TODO_DOC}

\item{logo}{TODO_DOC}

\item{theme}{TODO_DOC}

\item{tools}{TODO_DOC}

\item{toolbar_location}{Where the toolbar will be located. If set to None, no toolbar will be attached to the plot. One of 'above', 'below', 'left', 'right'.}

\item{toolbar_sticky}{Stick the toolbar to the edge of the plot. Default: TRUE If FALSE, the toolbar will be outside of the axes, titles etc.}

\item{h_symmetry}{Whether the total horizontal padding on both sides of the plot will be made equal (the left or right padding amount, whichever is larger).}

\item{v_symmetry}{Whether the total vertical padding on both sides of the plot will be made equal (the top or bottom padding amount, whichever is larger).}

\item{lod_factor}{Decimation factor to use when applying level-of-detail decimation.}

\item{lod_interval}{Interval (in ms) during which an interactive tool event will enable level-of-detail downsampling.}

\item{lod_threshold}{A number of data points, above which level-of-detail downsampling may be performed by glyph renderers. Set to ``None`` to disable any level-of-detail downsampling.}

\item{lod_timeout}{Timeout (in ms) for checking whether interactive tool events are still occurring. Once level-of-detail mode is enabled, a check is made every ``lod_timeout`` ms. If no interactive tool events have happened, level-of-detail mode is disabled.}

\item{output_backend}{Specify the output backend for the plot area. Default is "canvas", HTML5 Canvas. Note: When set to "webgl", glyphs without a WebGL rendering implementation will fall back to rendering onto 2D canvas. Must be one of 'canvas', 'svg', 'webgl'.}

\item{min_border}{A convenience property to set all all the ``min_border_X`` properties to the same value. If an individual border property is explicitly set, it will override ``min_border``.}

\item{min_border_left}{Minimum size in pixels of the padding region to the left of the central plot region. Note: This is a *minimum*. The padding region may expand as needed to accommodate titles or axes, etc.}

\item{min_border_bottom}{Minimum size in pixels of the padding region below the bottom of the central plot region. Note: This is a *minimum*. The padding region may expand as needed to accommodate titles or axes, etc.}

\item{min_border_right}{Minimum size in pixels of the padding region to the right of the central plot region. Note: This is a *minimum*. The padding region may expand as needed to accommodate titles or axes, etc.}

\item{min_border_top}{Minimum size in pixels of the padding region above the top of the central plot region. Note: This is a *minimum*. The padding region may expand as needed to accommodate titles or axes, etc.}

\item{hidpi}{Whether to use HiDPI mode when available.}
}
\description{
Initialize a Bokeh Google Map plot
}
\note{
This can be used in the same way as \code{\link{figure}}, adding layers on top of the Google Map.
There is an open issue documenting points appearing to sometimes be a few pixels off from their intended location.
Google has its own terms of service for using Google Maps API and any use of rbokeh with Google Maps must be within Google's Terms of Service
}
\seealso{
\code{\link{gmap_style}}
}
