% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer__utils.R
\name{grab}
\alias{grab}
\title{Grab function}
\usage{
grab(..., dots, nullData = FALSE)
}
\arguments{
\item{...}{arguments (that may or may not be able to evaluate right away) to be captured}

\item{dots}{output from \code{lazy_dots()} from the parent function}

\item{nullData}{boolen to determine if data args are being passed in}
}
\value{
list of three groups: data, info, and params
}
\description{
This function takes in all arguments and lazy wraps them.  Once wrapped, they are placed into one of three groups: data, info, or params.  params should be able to be sent to \code{check_opts} without issue
}

