% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer__utils.R
\name{grab_data_index}
\alias{grab_data_index}
\title{helper function to get all 'data' positions from the given names}
\usage{
grab_data_index(nameVals)
}
\description{
helper function to get all 'data' positions from the given names
}

