% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer__utils.R
\name{grab_param_index}
\alias{grab_param_index}
\title{helper function to get all 'params' positions from the given names}
\usage{
grab_param_index(dataIdx, regularIdx)
}
\description{
helper function to get all 'params' positions from the given names
}

