% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer__utils.R
\name{grab_regular_index}
\alias{grab_regular_index}
\title{helper function to get all 'info' positions from the given names}
\usage{
grab_regular_index(nameVals)
}
\description{
helper function to get all 'info' positions from the given names
}

