% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{jitter_cat}
\alias{jitter_cat}
\title{Add a small amount of (rbokeh-compatible) noise to a character vector}
\usage{
jitter_cat(x, factor = 0.5)
}
\arguments{
\item{x}{numeric vector to which jitter should be added}

\item{factor}{a factor between 0 and 1 that}
}
\description{
Add a small amount of (rbokeh-compatible) noise to a character vector
}
