% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbokeh-package.R
\docType{data}
\name{lesmis}
\alias{lesmis}
\title{Character co-occurence from Les Miserables}
\usage{
lesmis
}
\description{
Number of times two characters from Les Miserables appear in the same
scene(?). This dataset is used as an example to generate a heatmap-like
visualization in the main bokeh distribution
\href{http://bokeh.pydata.org/en/latest/docs/gallery/les_mis.html}{here}.
The data is stored in a list with \code{data.frames} \code{lesmis$nodes} and
\code{lesmis$links}
}
\examples{
\donttest{
data("lesmis", package="rbokeh")

links <- lesmis$links
nodes <- lesmis$nodes
names <- nodes[order(nodes$group, decreasing=TRUE),]$name

colormap <-  c("#444444", "#a6cee3", "#1f78b4", "#b2df8a", "#33a02c", "#fb9a99",
               "#e31a1c", "#fdbf6f", "#ff7f00", "#cab2d6", "#6a3d9a")

counts <- matrix(0, nrow=nrow(nodes), ncol=nrow(nodes),
                 dimnames=list(nodes$name, nodes$name))
for (i in 1:nrow(links)) {
  counts[links$source[i] + 1L, links$target[i] + 1L] <- links$value[i]
  counts[links$target[i] + 1L, links$source[i] + 1L] <- links$value[i]
}

linx <- reshape2::melt(counts) \%>\%
  setNames(c('xname', 'yname', 'count')) \%>\%
  transform(xname=as.character(xname), yname=as.character(yname, names)) \%>\%
  merge(nodes, by.x='xname', by.y='name') \%>\%
  merge(nodes, by.x='yname', by.y='name') \%>\%
  transform(xname=factor(xname, names), yname=factor(yname, names)) \%>\%
  transform(alphas=pmin(count / 4, 0.9) + 0.1,
            colors=ifelse(group.x == group.y, colormap[group.x + 1], '#3f3f3f'),
            xl=as.integer(xname), xr=as.integer(xname) + 0.9,
            yb=as.integer(yname), yt=as.integer(yname) + 0.9,
            stringsAsFactors=FALSE)

figure(linx, title="Les Mis Occurrences", width=800, height=800,
       xaxes='above') \%>\%
  ly_rect(xl, yb, xr, yt, color=colors, alpha=alphas,
          hover=c('xname', 'yname', 'count'),
          legend=FALSE)

}
}
\keyword{data}

