% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_base_plus.R
\name{ly_abline}
\alias{ly_abline}
\title{Add an "abline" layer to a Bokeh figure}
\usage{
ly_abline(fig, a = NULL, b = NULL, v = NULL, h = NULL, coef = NULL,
  color = "black", alpha = NULL, width = 1, type = 1,
  hov_color = NULL, hov_alpha = NULL, ns_color = NULL, ns_alpha = NULL,
  sel_color = NULL, sel_alpha = NULL, legend = NULL, lname = NULL,
  lgroup = NULL, ...)
}
\arguments{
\item{fig}{Figure to modify.}

\item{a, b}{the intercept and slope of the line(s) to draw}

\item{v}{the x value(s) for vertical lines}

\item{h}{the y value(s) for horizontal lines}

\item{coef}{a vector of length two giving the intercept and slope}
}
\description{
Draws one or more straight lines.
}
\examples{
# abline with mixed axes for h and v
figure() \%>\%
  ly_points(1:26, letters) \%>\%
  ly_abline(h = "j") \%>\%
  ly_abline(v = 10)

# multiple hv lines
figure() \%>\%
  ly_points(1:10) \%>\%
  ly_abline(v = 1:10) \%>\%
  ly_abline(h = 1:10)

# multiple ab lines
figure() \%>\%
  ly_points(0:10) \%>\%
  ly_abline(0, seq(0, 1, by = 0.1))
}
\seealso{
Other layer functions: \code{\link{ly_annular_wedge}},
  \code{\link{ly_annulus}}, \code{\link{ly_arc}},
  \code{\link{ly_bar}}, \code{\link{ly_bezier}},
  \code{\link{ly_boxplot}}, \code{\link{ly_contour}},
  \code{\link{ly_crect}}, \code{\link{ly_curve}},
  \code{\link{ly_density}}, \code{\link{ly_hist}},
  \code{\link{ly_image_url}}, \code{\link{ly_image}},
  \code{\link{ly_lines}}, \code{\link{ly_multi_line}},
  \code{\link{ly_oval}}, \code{\link{ly_patch}},
  \code{\link{ly_points}}, \code{\link{ly_polygons}},
  \code{\link{ly_quadratic}}, \code{\link{ly_quantile}},
  \code{\link{ly_ray}}, \code{\link{ly_rect}},
  \code{\link{ly_segments}}, \code{\link{ly_text}},
  \code{\link{ly_wedge}}
}
