% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/layer_glyphs.R
\name{ly_annular_wedge}
\alias{ly_annular_wedge}
\title{Add an "annular_wedge" layer to a Bokeh figure}
\usage{
ly_annular_wedge(fig, x, y = NULL, data = NULL, inner_radius = 0.1,
  outer_radius = 0.3, start_angle = 0, end_angle = 2 * pi,
  direction = "anticlock", color = NULL, alpha = NULL,
  fill_color = NULL, fill_alpha = 0.75, line_color = NULL,
  line_width = 1, line_alpha = NULL, hover = NULL, legend = NULL,
  lname = NULL, lgroup = NULL, ...)
}
\arguments{
\item{x}{the x-coordinates of the centers of the annular wedges}

\item{y}{the y-coordinates of the centers of the annular wedges}

\item{data}{an optional data frame, providing the source for inputs x, y, and other glyph properties}

\item{inner_radius}{the inner radii of the annular wedges}

\item{outer_radius}{The outer radii of the annular wedges}

\item{start_angle}{the angles to start the annular wedges, in radians, as measured from the horizontal}

\item{end_angle}{the angles to end the annular wedges, in radians, as measured from the horizontal}

\item{direction}{which direction to stroke between the start and end angles ("clock", "anticlock")}

\item{color}{color for the glyph - a hex code (with no alpha) or any of the 147 named CSS colors, e.g 'green', 'indigo' - for glyphs with both fill and line properties, see "Handling color" below}

\item{alpha}{the alpha transparency of the glyph between 0 (transparent) and 1 (opaque) - if glyph has both fill and color properties, see "Handling alpha" below}

\item{fill_color}{color to use to fill the glyph with - a hex code (with no alpha) or any of the 147 named CSS colors, e.g 'green', 'indigo'}

\item{fill_alpha}{transparency value between 0 (transparent) and 1 (opaque)}

\item{line_color}{color to use to stroke lines with - a hex code (with no alpha) or any of the 147 named CSS colors, e.g 'green', 'indigo'}

\item{line_width}{stroke width in units of pixels}

\item{line_alpha}{transparency value between 0 (transparent) and 1 (opaque)}

\item{hover}{a data frame of variables to be displayed when hovering over the glyph or a vector of variable names to be found in the \code{data} argument}

\item{legend}{either a logical specifying not to plot a legend for this layer (FALSE) or a string indicating the name of the legend entry for this layer (note that when mapping plot attributes to variables in \code{data}, a legend is automatically created and does not need to be specified)}

\item{lname}{}

\item{lgroup}{}

\item{\ldots}{additional parameters}
}
\description{
Add an "annular_wedge" layer to a Bokeh figure
}
\section{Handling color}{
 The \code{color} parameter is a high-level plot attribute that provides default behavior for coloring glyphs.
\itemize{
  \item When using a glyph that only has line properties, this will be the color of the line.
  \item When using a glyph that has has line and fill properties, this will be the color of the line and the fill, with the alpha level of the fill reduced by 50\%.
  \item If full control over fill and line color is desired, the \code{fill_color} and \code{line_color} attributes can be specified explicitly and will override \code{color}.
}
}

\section{Handling alpha}{
 The \code{alpha} is a high-level plot attribute that sets the transparency of the glyph being plotted.
\itemize{
  \item When using a glyph that only has line properties, this will be the alpha of the line.
  \item When using a glyph that has has line and fill properties, this will be the alpha of the line and the alpha of the fill will be set to 50\% of this value.
  \item Individual fill and line alpha can be specified with \code{fill_alpha} and \code{line_alpha} and will override \code{alpha}.
}
}

