% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_stat.R
\name{ly_bar}
\alias{ly_bar}
\title{Add a "barchart" layer to a Bokeh figure}
\usage{
ly_bar(fig, x = NULL, y = NULL, data = figure_data(fig), color = NULL,
  alpha = 1, position = c("stack", "fill", "dodge"), width = 0.9,
  hover = FALSE, origin = NULL, breaks = NULL, right = FALSE,
  binwidth = NULL, lname = NULL, lgroup = NULL, legend = TRUE, ...)
}
\arguments{
\item{fig}{Figure to modify.}

\item{x}{values or field name for x variable, or if NULL, x-axis will be counts of y}

\item{y}{values or field name for y variable, or if NULL, y-axis will be counts of x}

\item{data}{an optional data frame, providing the source for inputs x, y, and color properties}

\item{position}{either "stack", "fill", or "dodge" (see details)}

\item{width}{with of each bar, a value between 0 (no width) and 1 (full width)}

\item{hover}{logical - should a hover tool be added to show the value of each bar?}

\item{origin, breaks, right, binwidth}{parameters to be used for binning x when it is continuous (not yet implemented)}
}
\description{
Draws a bar chart
}
\details{
This function expects one of either x or y to be categorical and the other to be numeric or NULL.  The numeric variable is summed for each categorical variable and bars are plotted.  If no numeric variable is supplied, the unique values of the categorical variable will be tabulated.  Within each categorical variable, if color maps to another grouping variable then the bars are split up.  In this case, there are three ways to display the bars with the \code{position} argument.  The default, "stack" will stack the bars.  The "fill" choice will show the relative proportion for each group within each categorical variable level, stacking the bars.  The "dodge" choice will plot the bars for each level of the categorical variable side by side.
}
\seealso{
Other layer functions: \code{\link{ly_annular_wedge}},
  \code{\link{ly_annulus}}, \code{\link{ly_arc}},
  \code{\link{ly_bezier}}, \code{\link{ly_boxplot}},
  \code{\link{ly_contour}}, \code{\link{ly_crect}},
  \code{\link{ly_curve}}, \code{\link{ly_density}},
  \code{\link{ly_hist}}, \code{\link{ly_image_url}},
  \code{\link{ly_image}}, \code{\link{ly_lines}},
  \code{\link{ly_multi_line}}, \code{\link{ly_oval}},
  \code{\link{ly_patch}}, \code{\link{ly_points}},
  \code{\link{ly_polygons}}, \code{\link{ly_quadratic}},
  \code{\link{ly_quantile}}, \code{\link{ly_ray}},
  \code{\link{ly_rect}}, \code{\link{ly_segments}},
  \code{\link{ly_text}}, \code{\link{ly_wedge}}
}
