% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_stat.R
\name{ly_boxplot}
\alias{ly_boxplot}
\title{Add a "boxplot" layer to a Bokeh figure}
\usage{
ly_boxplot(fig, x, y = NULL, data = figure_data(fig), width = 0.9,
  coef = 1.5, color = "blue", alpha = 1, hov_color = NULL,
  hov_alpha = NULL, ns_color = NULL, ns_alpha = NULL, sel_color = NULL,
  sel_alpha = NULL, outlier_glyph = 1, outlier_size = 10, lname = NULL,
  lgroup = NULL, ...)
}
\arguments{
\item{fig}{Figure to modify.}

\item{x}{either a numeric vector or a factor}

\item{y}{either a numeric vector or a factor}

\item{data}{an optional data frame, providing the source for x and y}

\item{width}{with of each box, a value between 0 (no width) and 1 (full width)}

\item{coef}{see \code{\link[grDevices]{boxplot.stats}}}

\item{outlier_glyph}{the glyph used to plot the outliers. If set to
\code{NA}, no outlier points are plotted. Run \code{point_types()} for
possible values.}

\item{outlier_size}{the size of the glyph used to plot outliers. If set to
\code{NA}, no outlier points are plotted.}
}
\description{
Add a "boxplot" layer to a Bokeh figure
}
\examples{
figure(ylab = "Height (inches)", width = 600) \%>\%
  ly_boxplot(voice.part, height, data = lattice::singer)
}
\seealso{
Other layer functions: \code{\link{ly_abline}},
  \code{\link{ly_annular_wedge}}, \code{\link{ly_annulus}},
  \code{\link{ly_arc}}, \code{\link{ly_bar}},
  \code{\link{ly_bezier}}, \code{\link{ly_contour}},
  \code{\link{ly_crect}}, \code{\link{ly_curve}},
  \code{\link{ly_density}}, \code{\link{ly_hist}},
  \code{\link{ly_image_url}}, \code{\link{ly_image}},
  \code{\link{ly_lines}}, \code{\link{ly_multi_line}},
  \code{\link{ly_oval}}, \code{\link{ly_patch}},
  \code{\link{ly_points}}, \code{\link{ly_polygons}},
  \code{\link{ly_quadratic}}, \code{\link{ly_quantile}},
  \code{\link{ly_ray}}, \code{\link{ly_rect}},
  \code{\link{ly_segments}}, \code{\link{ly_text}},
  \code{\link{ly_wedge}}
}
