% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_base_plus.R
\name{ly_curve}
\alias{ly_curve}
\title{Add a "curve" layer to a Bokeh figure}
\usage{
ly_curve(fig, expr, from = NULL, to = NULL, n = 101, width = 1,
  type = 1, color = "black", alpha = 1, hov_color = NULL,
  hov_alpha = NULL, ns_color = NULL, ns_alpha = NULL, legend = NULL,
  lname = NULL, lgroup = NULL, ...)
}
\arguments{
\item{fig}{Figure to modify.}

\item{expr, from, to, n}{parameters sent to \code{\link[graphics]{curve}}}
}
\description{
Draws a curve corresponding to a function over the interval \code{[from, to]}.
}
\examples{
\donttest{
chippy <- function(x) sin(cos(x)*exp(-x/2))
figure(width = 800) \%>\%
  ly_curve(chippy, -8, 7, n = 2001)
}
}
\seealso{
Other layer functions: \code{\link{ly_annular_wedge}},
  \code{\link{ly_annulus}}, \code{\link{ly_arc}},
  \code{\link{ly_bar}}, \code{\link{ly_bezier}},
  \code{\link{ly_boxplot}}, \code{\link{ly_contour}},
  \code{\link{ly_crect}}, \code{\link{ly_density}},
  \code{\link{ly_hist}}, \code{\link{ly_image_url}},
  \code{\link{ly_image}}, \code{\link{ly_lines}},
  \code{\link{ly_multi_line}}, \code{\link{ly_oval}},
  \code{\link{ly_patch}}, \code{\link{ly_points}},
  \code{\link{ly_polygons}}, \code{\link{ly_quadratic}},
  \code{\link{ly_quantile}}, \code{\link{ly_ray}},
  \code{\link{ly_rect}}, \code{\link{ly_segments}},
  \code{\link{ly_text}}, \code{\link{ly_wedge}}
}
