% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_stat.R
\name{ly_hist}
\alias{ly_hist}
\title{Add a "hist" layer to a Bokeh figure}
\usage{
ly_hist(fig, x, data = figure_data(fig), breaks = "Sturges", freq = TRUE,
  include.lowest = TRUE, right = TRUE, color = NULL, alpha = 1,
  hov_color = NULL, hov_alpha = NULL, ns_color = NULL, ns_alpha = NULL,
  sel_color = NULL, sel_alpha = NULL, lname = NULL, lgroup = NULL, ...)
}
\arguments{
\item{fig}{Figure to modify.}

\item{x}{either a vector to be passed to \code{\link[graphics]{hist}} or an object of class "histogram"}

\item{data}{an optional data frame, providing the source for x}

\item{breaks, freq, include.lowest, right}{parameters passed to \code{\link[graphics]{hist}}}
}
\description{
Draws a histogram
}
\examples{
h <- figure(width = 600, height = 400) \%>\%
  ly_hist(eruptions, data = faithful, breaks = 40, freq = FALSE) \%>\%
  ly_density(eruptions, data = faithful)
h
}
\seealso{
Other layer functions: \code{\link{ly_annular_wedge}},
  \code{\link{ly_annulus}}, \code{\link{ly_arc}},
  \code{\link{ly_bar}}, \code{\link{ly_bezier}},
  \code{\link{ly_boxplot}}, \code{\link{ly_contour}},
  \code{\link{ly_crect}}, \code{\link{ly_curve}},
  \code{\link{ly_density}}, \code{\link{ly_image_url}},
  \code{\link{ly_image}}, \code{\link{ly_lines}},
  \code{\link{ly_multi_line}}, \code{\link{ly_oval}},
  \code{\link{ly_patch}}, \code{\link{ly_points}},
  \code{\link{ly_polygons}}, \code{\link{ly_quadratic}},
  \code{\link{ly_quantile}}, \code{\link{ly_ray}},
  \code{\link{ly_rect}}, \code{\link{ly_segments}},
  \code{\link{ly_text}}, \code{\link{ly_wedge}}
}
