% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barebones.R
\name{ly_line}
\alias{ly_line}
\title{Bare-bones line layer function}
\usage{
ly_line(obj, x, y = x)
}
\arguments{
\item{obj}{an rbokeh object}

\item{x}{x values}

\item{y}{y values}
}
\description{
Bare-bones line layer function
}
