% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_base.R
\name{ly_multi_line}
\alias{ly_multi_line}
\title{Add a "multi_line" layer to a Bokeh figure}
\usage{
ly_multi_line(fig, xs = NULL, ys = NULL, data = figure_data(fig),
  type = 1, width = 1, color = NULL, alpha = NULL, hover = NULL,
  legend = TRUE, hov_color = NULL, hov_alpha = NULL, ns_color = NULL,
  ns_alpha = NULL, lgroup = NULL, lname = NULL, ...)
}
\arguments{
\item{fig}{figure to modify}

\item{xs}{list of vectors of x coordinates}

\item{ys}{list of vectors of y coordinates}
}
\description{
Draws multiple lines with the given lists of coordinates.
}
\seealso{
Other layer functions: \code{\link{ly_abline}},
  \code{\link{ly_annular_wedge}}, \code{\link{ly_annulus}},
  \code{\link{ly_arc}}, \code{\link{ly_bar}},
  \code{\link{ly_bezier}}, \code{\link{ly_boxplot}},
  \code{\link{ly_contour}}, \code{\link{ly_crect}},
  \code{\link{ly_curve}}, \code{\link{ly_density}},
  \code{\link{ly_hist}}, \code{\link{ly_image_url}},
  \code{\link{ly_image}}, \code{\link{ly_lines}},
  \code{\link{ly_oval}}, \code{\link{ly_patch}},
  \code{\link{ly_points}}, \code{\link{ly_polygons}},
  \code{\link{ly_quadratic}}, \code{\link{ly_quantile}},
  \code{\link{ly_ray}}, \code{\link{ly_rect}},
  \code{\link{ly_segments}}, \code{\link{ly_text}},
  \code{\link{ly_wedge}}
}
