% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_base.R
\name{ly_points}
\alias{ly_points}
\title{Add a "points" layer to a Bokeh figure
Draws points with the given coordinates.}
\usage{
ly_points(fig, x = NULL, y = NULL, data = figure_data(fig), glyph = 21,
  color = NULL, alpha = NULL, size = NULL, hover = NULL,
  legend = TRUE, hov_color = NULL, hov_alpha = NULL, ns_color = NULL,
  ns_alpha = NULL, lgroup = NULL, lname = NULL, ...)
}
\arguments{
\item{fig}{Figure to modify.}

\item{x}{values or field name of line x coordinates}

\item{y}{values or field name of line y coordinates}

\item{data}{an optional data frame, providing the source for inputs x, y, and other glyph properties}
}
\description{
Add a "points" layer to a Bokeh figure
Draws points with the given coordinates.
}
\seealso{
Other layer functions: \code{\link{ly_abline}},
  \code{\link{ly_annular_wedge}}, \code{\link{ly_annulus}},
  \code{\link{ly_arc}}, \code{\link{ly_bar}},
  \code{\link{ly_bezier}}, \code{\link{ly_boxplot}},
  \code{\link{ly_contour}}, \code{\link{ly_crect}},
  \code{\link{ly_curve}}, \code{\link{ly_density}},
  \code{\link{ly_hist}}, \code{\link{ly_image_url}},
  \code{\link{ly_image}}, \code{\link{ly_lines}},
  \code{\link{ly_multi_line}}, \code{\link{ly_oval}},
  \code{\link{ly_patch}}, \code{\link{ly_polygons}},
  \code{\link{ly_quadratic}}, \code{\link{ly_quantile}},
  \code{\link{ly_ray}}, \code{\link{ly_rect}},
  \code{\link{ly_segments}}, \code{\link{ly_text}},
  \code{\link{ly_wedge}}
}
