% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_base.R
\name{ly_quadratic}
\alias{ly_quadratic}
\title{Add a "quadratic" layer to a Bokeh figure}
\usage{
ly_quadratic(fig, x0, y0, x1, y1, cx, cy, data = figure_data(fig),
  color = NULL, alpha = NULL, type = 1, width = 1, hov_color = NULL,
  hov_alpha = NULL, ns_color = NULL, ns_alpha = NULL, sel_color = NULL,
  sel_alpha = NULL, hover = NULL, url = NULL, legend = TRUE,
  lname = NULL, lgroup = NULL, ...)
}
\arguments{
\item{fig}{figure to modify}

\item{x0}{values or field name of starting x coordinates}

\item{y0}{values or field name of starting y coordinates}

\item{x1}{values or field name of ending x coordinates}

\item{y1}{values or field name of ending y coordinates}

\item{cx}{values or field name of control point x coordinates}

\item{cy}{values or field name of control point y coordinates}

\item{data}{an optional data frame, providing the source for start, end, and control point intputs, as well as other glyph properties}
}
\description{
Draws quadratic curves with the given starting, ending, and control points.
}
\seealso{
Other layer functions: \code{\link{ly_abline}},
  \code{\link{ly_annular_wedge}}, \code{\link{ly_annulus}},
  \code{\link{ly_arc}}, \code{\link{ly_bar}},
  \code{\link{ly_bezier}}, \code{\link{ly_boxplot}},
  \code{\link{ly_contour}}, \code{\link{ly_crect}},
  \code{\link{ly_curve}}, \code{\link{ly_density}},
  \code{\link{ly_hist}}, \code{\link{ly_image_url}},
  \code{\link{ly_image}}, \code{\link{ly_lines}},
  \code{\link{ly_multi_line}}, \code{\link{ly_oval}},
  \code{\link{ly_patch}}, \code{\link{ly_points}},
  \code{\link{ly_polygons}}, \code{\link{ly_quantile}},
  \code{\link{ly_ray}}, \code{\link{ly_rect}},
  \code{\link{ly_segments}}, \code{\link{ly_text}},
  \code{\link{ly_wedge}}
}
