% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_stat.R
\name{ly_quantile}
\alias{ly_quantile}
\title{Add a "quantile" layer to a Bokeh figure}
\usage{
ly_quantile(fig, x, group = NULL, data = figure_data(fig), probs = NULL,
  distn = stats::qunif, ncutoff = 200, color = NULL, alpha = 1,
  hov_color = NULL, hov_alpha = NULL, ns_color = NULL, ns_alpha = NULL,
  sel_color = NULL, sel_alpha = NULL, legend = TRUE, lname = NULL,
  lgroup = NULL, ...)
}
\arguments{
\item{fig}{Figure to modify.}

\item{x}{numeric vector or field name of variable to compute sample quantiles for}

\item{group}{values or field name of a grouping variable to break quantile computations up by}

\item{data}{an optional data frame, providing the source for x}

\item{probs}{numeric vector of probabilities with values in \code{[0,1]} at which to compute quantiles - if \code{NULL}, every point of \code{x} is a quantile}

\item{distn}{quantile function to use on the x-axis (e.g. \code{\link[stats]{qnorm}}) - default is \code{\link[stats]{qunif}},}

\item{ncutoff}{if the length of \code{x} exceeds this value and \code{probs} is not specified, compute quantiles at \code{ncutoff} points}
}
\description{
Draws quantiles
}
\examples{
figure(legend_location = "top_left") \%>\%
  ly_quantile(Sepal.Length, group = Species, data = iris)
}
\seealso{
Other layer functions: \code{\link{ly_abline}},
  \code{\link{ly_annular_wedge}}, \code{\link{ly_annulus}},
  \code{\link{ly_arc}}, \code{\link{ly_bar}},
  \code{\link{ly_bezier}}, \code{\link{ly_boxplot}},
  \code{\link{ly_contour}}, \code{\link{ly_crect}},
  \code{\link{ly_curve}}, \code{\link{ly_density}},
  \code{\link{ly_hist}}, \code{\link{ly_image_url}},
  \code{\link{ly_image}}, \code{\link{ly_lines}},
  \code{\link{ly_multi_line}}, \code{\link{ly_oval}},
  \code{\link{ly_patch}}, \code{\link{ly_points}},
  \code{\link{ly_polygons}}, \code{\link{ly_quadratic}},
  \code{\link{ly_ray}}, \code{\link{ly_rect}},
  \code{\link{ly_segments}}, \code{\link{ly_text}},
  \code{\link{ly_wedge}}
}
