% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_base.R
\name{ly_ray}
\alias{ly_ray}
\title{Add a "ray" layer to a Bokeh figure}
\usage{
ly_ray(fig, x, y = NULL, data = figure_data(fig), length = NULL,
  angle = NULL, type = 1, width = 1, color = NULL, alpha = NULL,
  hover = NULL, url = NULL, hov_color = NULL, hov_alpha = NULL,
  ns_color = NULL, ns_alpha = NULL, sel_color = NULL, sel_alpha = NULL,
  legend = NULL, lname = NULL, lgroup = NULL, visible = TRUE, ...)
}
\arguments{
\item{fig}{figure to modify}

\item{x}{values or field name of center x coordinates}

\item{y}{values or field name of center y coordinates}

\item{data}{an optional data frame, providing the source for inputs x, y, and other glyph properties}

\item{length}{values or field name of ray lengths in screen units}

\item{angle}{values or field name of ray angles}
}
\description{
Draws line segments starting at the given coordinate and extending the given length at the given angle.
}
\seealso{
Other layer functions: \code{\link{ly_abline}},
  \code{\link{ly_annular_wedge}}, \code{\link{ly_annulus}},
  \code{\link{ly_arc}}, \code{\link{ly_bar}},
  \code{\link{ly_bezier}}, \code{\link{ly_boxplot}},
  \code{\link{ly_contour}}, \code{\link{ly_crect}},
  \code{\link{ly_curve}}, \code{\link{ly_density}},
  \code{\link{ly_hist}}, \code{\link{ly_image_url}},
  \code{\link{ly_image}}, \code{\link{ly_lines}},
  \code{\link{ly_multi_line}}, \code{\link{ly_oval}},
  \code{\link{ly_patch}}, \code{\link{ly_points}},
  \code{\link{ly_polygons}}, \code{\link{ly_quadratic}},
  \code{\link{ly_quantile}}, \code{\link{ly_rect}},
  \code{\link{ly_segments}}, \code{\link{ly_text}},
  \code{\link{ly_wedge}}
}
