% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_base.R
\name{ly_segments}
\alias{ly_segments}
\title{Add a "segments" layer to a Bokeh figure}
\usage{
ly_segments(fig, x0, y0, x1, y1, data = figure_data(fig), type = 1,
  width = 1, color = NULL, alpha = NULL, hov_color = NULL,
  hov_alpha = NULL, ns_color = NULL, ns_alpha = NULL, sel_color = NULL,
  sel_alpha = NULL, hover = NULL, url = NULL, legend = TRUE,
  lname = NULL, lgroup = NULL, ...)
}
\arguments{
\item{fig}{Figure to modify.}

\item{x0}{values or field name of starting x coordinates}

\item{y0}{values or field name of starting y coordinates}

\item{x1}{values or field name of ending x coordinates}

\item{y1}{values or field name of ending y coordinates}

\item{data}{an optional data frame, providing the source for inputs x, y, and other glyph properties}
}
\description{
Draws line segments with the given starting and ending coordinates.
}
\seealso{
Other layer functions: \code{\link{ly_annular_wedge}},
  \code{\link{ly_annulus}}, \code{\link{ly_arc}},
  \code{\link{ly_bar}}, \code{\link{ly_bezier}},
  \code{\link{ly_boxplot}}, \code{\link{ly_contour}},
  \code{\link{ly_crect}}, \code{\link{ly_curve}},
  \code{\link{ly_density}}, \code{\link{ly_hist}},
  \code{\link{ly_image_url}}, \code{\link{ly_image}},
  \code{\link{ly_lines}}, \code{\link{ly_multi_line}},
  \code{\link{ly_oval}}, \code{\link{ly_patch}},
  \code{\link{ly_points}}, \code{\link{ly_polygons}},
  \code{\link{ly_quadratic}}, \code{\link{ly_quantile}},
  \code{\link{ly_ray}}, \code{\link{ly_rect}},
  \code{\link{ly_text}}, \code{\link{ly_wedge}}
}
