% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_base.R
\name{ly_wedge}
\alias{ly_wedge}
\title{Add a "wedge" layer to a Bokeh figure}
\usage{
ly_wedge(fig, x, y = NULL, data = figure_data(fig), radius = 0.2,
  start_angle = 0, end_angle = 2 * pi, direction = "anticlock",
  color = NULL, alpha = NULL, hov_color = NULL, hov_alpha = NULL,
  ns_color = NULL, ns_alpha = NULL, sel_color = NULL, sel_alpha = NULL,
  hover = NULL, url = NULL, legend = TRUE, lname = NULL,
  lgroup = NULL, ...)
}
\arguments{
\item{fig}{Figure to modify.}

\item{x}{values or field name of center x coordinates}

\item{y}{values or field name of center y coordinates}

\item{data}{an optional data frame, providing the source for inputs x, y, and other glyph properties}

\item{radius}{values or field name of wedge radii}

\item{start_angle}{the angles to start the wedges, in radians, as measured from the horizontal}

\item{end_angle}{the angles to end the wedges, in radians, as measured from the horizontal}

\item{direction}{direction to turn between starting and ending angles ("anticlock", "clock")}
}
\description{
Add a "wedge" layer to a Bokeh figure
}
\examples{
rescale <- function(x)
  (x - min(x)) / diff(range(x))
figure() \%>\%
  ly_wedge(Sepal.Length, Sepal.Width, data = iris,
    end_angle = rescale(Petal.Length) * 2 * pi, color = Species,
    radius = 0.15, alpha = 0.5,
    hover = Species)
}
\seealso{
Other layer functions: \code{\link{ly_annular_wedge}},
  \code{\link{ly_annulus}}, \code{\link{ly_arc}},
  \code{\link{ly_bar}}, \code{\link{ly_bezier}},
  \code{\link{ly_boxplot}}, \code{\link{ly_contour}},
  \code{\link{ly_crect}}, \code{\link{ly_curve}},
  \code{\link{ly_density}}, \code{\link{ly_hist}},
  \code{\link{ly_image_url}}, \code{\link{ly_image}},
  \code{\link{ly_lines}}, \code{\link{ly_multi_line}},
  \code{\link{ly_oval}}, \code{\link{ly_patch}},
  \code{\link{ly_points}}, \code{\link{ly_polygons}},
  \code{\link{ly_quadratic}}, \code{\link{ly_quantile}},
  \code{\link{ly_ray}}, \code{\link{ly_rect}},
  \code{\link{ly_segments}}, \code{\link{ly_text}}
}
