% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/point_types.R
\name{point_types}
\alias{point_types}
\title{Display glyph types available for ly_points()}
\usage{
point_types(size = 25, color = "blue")
}
\arguments{
\item{size}{size of the glyph}

\item{color}{color to use for line and fill properties}
}
\description{
Display glyph types available for ly_points()
}

