% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.BokehFigure}
\alias{print.BokehFigure}
\title{Print a Bokeh Figure}
\usage{
\method{print}{BokehFigure}(x, ...)
}
\arguments{
\item{x}{Bokeh figure}

\item{\ldots}{further arguments - most importantly \code{debug = TRUE}
will print information about the figure in the javascript console in the
web browser}
}
\description{
Print a Bokeh Figure
}

