% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print.R
\name{print.BokehGridPlot}
\alias{print.BokehGridPlot}
\title{Print a Bokeh grid plot}
\usage{
\method{print}{BokehGridPlot}(x, ...)
}
\arguments{
\item{x}{BokehGridPlot object}

\item{\ldots}{further arguments - most importantly \code{debug = TRUE}
will print information about the figure in the javascript console in the
web browser}
}
\description{
Print a Bokeh grid plot
}

