% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/_doc.R
\docType{package}
\name{rbokeh-package}
\alias{rbokeh}
\alias{rbokeh-package}
\title{rbokeh: R interface for Bokeh}
\description{
R interface for creating plots in Bokeh.  Bokeh by Continuum Analytics, \url{http://bokeh.pydata.org/en/latest/}
}
\details{
For full documentation on the package, visit \url{http://hafen.github.io/rbokeh}
}

