% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/json_helpers.R
\name{rbokeh2html}
\alias{rbokeh2html}
\title{Get the HTML content required to embed a Bokeh figure}
\usage{
rbokeh2html(fig)
}
\arguments{
\item{fig}{figure}
}
\description{
Get the HTML content required to embed a Bokeh figure
}

