% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{renderRbokeh}
\alias{renderRbokeh}
\title{Widget render function for use in Shiny}
\usage{
renderRbokeh(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{an expression that generates a rbokeh figure}

\item{env}{the environment in which to evaluate expr.}

\item{quoted}{is expr a quoted expression (with quote())? This is useful if you want to save an expression in a variable.}
}
\description{
Widget render function for use in Shiny
}
\seealso{
\code{\link{rbokehOutput}} for an example in Shiny
}
