% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/save_figure.R
\name{save_figure}
\alias{save_figure}
\title{Make a static png file for an rbokeh figure}
\usage{
save_figure(p, file, timeout = 500)
}
\arguments{
\item{p}{Bokeh figure object}

\item{file}{where to save png file}

\item{timeout}{plot render timeout in milliseconds (see details)}
}
\description{
Make a static png file for an rbokeh figure
}
\details{
This uses phantomjs (\url{http://phantomjs.org}) to render your rbokeh figure in a headless browser and take a screenshot of it, creating a static output.  This assumes that phantomjs has been installed on your machine and is available as a system call.  For plots that take longer to load and render, you may need to increase the value of \code{timeout}.  Note that this function is experimental.
}
\examples{
\donttest{
figure(tools = NULL) \%>\%
  ly_points(1:10) \%>\%
  save_figure("/tmp/test.png")
}
}

