% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/themes_constants.R
\name{scale_tableau_pal}
\alias{scale_continuous_pal}
\alias{scale_glyph}
\alias{scale_line_dash}
\alias{scale_line_width}
\alias{scale_size}
\alias{scale_tableau_pal}
\title{Scales for themes}
\usage{
scale_tableau_pal(pal = "Tableau10")

scale_glyph()

scale_continuous_pal()

scale_size(min = 2, max = 20)

scale_line_dash()

scale_line_width()
}
\arguments{
\item{pal}{palette name}

\item{min}{minimum value}

\item{max}{maximum value}
}
\description{
Scales for themes
}

