% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{set_theme}
\alias{set_theme}
\title{Set the theme for a figure}
\usage{
set_theme(fig, theme)
}
\arguments{
\item{fig}{a figure to set the theme for}

\item{theme}{theme}
}
\description{
Set the theme for a figure
}

