% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer__utils.R
\name{sub_names}
\alias{sub_names}
\title{retrieve and properly parse all data}
\usage{
sub_names(fig, data, argObj, processDataAndNames = TRUE)
}
\arguments{
\item{fig}{figure to be used}

\item{data}{data to be used}

\item{argObj}{args object supplied by \code{grab}}

\item{processDataAndNames}{boolean to determine if the data and xName and yName should be post processed}
}
\value{
list of three groups: data, info, and params.
}
\description{
retrieve and properly parse all data
}

