% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer__utils.R
\name{subset_arg_obj}
\alias{subset_arg_obj}
\title{Subset args object}
\usage{
subset_arg_obj(arg_obj, idxs)
}
\arguments{
\item{arg_obj}{args object to be subsetted}

\item{idxs}{indicies that should be used}
}
\value{
similar arg_obj, just subsetted at the suggested indicies
}
\description{
Subset args object
}

