% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_grid}
\alias{theme_grid}
\title{Override theme parameters for grid attributes}
\usage{
theme_grid(fig, which = c("x", "y"), band_fill_alpha = 1,
  band_fill_color = "gray", grid_line_alpha = 1, grid_line_cap = "butt",
  grid_line_color = "black", grid_line_dash = NULL,
  grid_line_dash_offset = 0, grid_line_join = "miter",
  grid_line_width = 1, minor_grid_line_alpha = 1,
  minor_grid_line_cap = "butt", minor_grid_line_color = "black",
  minor_grid_line_dash = NULL, minor_grid_line_dash_offset = 0,
  minor_grid_line_join = "miter", minor_grid_line_width = 1)
}
\arguments{
\item{fig}{Figure to modify.}

\item{which}{which grids to apply attributes to ("x" and/or "y")}

\item{band_fill_alpha}{The fill alpha of alternating bands between Grid lines.}

\item{band_fill_color}{The fill color of alternating bands between Grid lines.}

\item{grid_line_alpha}{The line alpha of the Grid lines.}

\item{grid_line_cap}{('butt', 'round', 'square') The line cap of the Grid lines.}

\item{grid_line_color}{The line color of the Grid lines.}

\item{grid_line_dash}{The line dash of the Grid lines.}

\item{grid_line_dash_offset}{The line dash offset of the Grid lines.}

\item{grid_line_join}{('miter', 'round', 'bevel') The line join of the Grid lines.}

\item{grid_line_width}{The line width of the Grid lines.}

\item{minor_grid_line_alpha}{The line alpha of the minor Grid lines.}

\item{minor_grid_line_cap}{('butt', 'round', 'square') The line cap of the minor Grid lines.}

\item{minor_grid_line_color}{The line color of the minor Grid lines.}

\item{minor_grid_line_dash}{The line dash of the minor Grid lines.}

\item{minor_grid_line_dash_offset}{The line dash offset of the minor Grid lines.}

\item{minor_grid_line_join}{('miter', 'round', 'bevel') The line join of the minor Grid lines.}

\item{minor_grid_line_width}{The line width of the minor Grid lines.}
}
\description{
Override theme parameters for grid attributes
}
