% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{bk_default_theme}
\alias{bk_default_theme}
\alias{bk_ggplot_theme}
\alias{bk_dark_theme}
\title{Themes}
\usage{
bk_default_theme()

bk_ggplot_theme()

bk_dark_theme()
}
\description{
Themes

Themes

Themes
}
