% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_axis_grid.R
\name{ticker_cat}
\alias{ticker_cat}
\title{Generate ticks for categorical ranges.}
\usage{
ticker_cat(clear = FALSE)
}
\arguments{
\item{clear}{A logical indicating whether all non-specified parameters should be cleared out of any exising specification in the figure. This can be useful when modifying a previously-defined figure.}
}
\value{
A specification that is used as the \code{ticker} argument for \code{\link{x_axis}} or \code{\link{y_axis}}.
}
\description{
Generate ticks for categorical ranges.
}
\note{
This function should never need to be used as there are no parameters for a categorical ticker.
}
