% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_axis_grid.R
\name{ticker_date}
\alias{ticker_date}
\title{Generate nice ticks across different date and time scales.}
\usage{
ticker_date(desired_num_ticks = 6L, num_minor_ticks = 0L, clear = FALSE)
}
\arguments{
\item{desired_num_ticks}{A desired target number of major tick positions to generate across the plot range. Note: This value is a suggestion, and ticker subclasses may ignore it entirely, or use it only as an ideal goal to approach as well as can be, in the context of a specific ticking strategy.}

\item{num_minor_ticks}{The number of minor tick positions to generate between adjacent major tick values.}

\item{clear}{A logical indicating whether all non-specified parameters should be cleared out of any exising specification in the figure. This can be useful when modifying a previously-defined figure.}
}
\value{
A specification that is used as the \code{ticker} argument for \code{\link{x_axis}} or \code{\link{y_axis}}.
}
\description{
Generate nice ticks across different date and time scales.
}
\examples{
figure(data = d, width = 1000) \%>\%
  ly_lines(date, co2_df) \%>\%
  x_axis(ticker = ticker_date(desired_num_ticks = 20))

figure(data = d, width = 1000) \%>\%
  ly_lines(date, co2_df) \%>\%
  x_axis(tickformatter = tickformatter_date(year = "'\%y"))

figure() \%>\%
  ly_points(Sys.time() - c(0:9) * 900, rnorm(10)) \%>\%
  x_axis(tickformatter = tickformatter_date(hourmin = "\%r \%Z"))
}
