% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_axis_grid.R
\name{ticker_fixed}
\alias{ticker_fixed}
\title{Generate ticks at fixed, explicitly supplied locations.}
\usage{
ticker_fixed(ticks = NULL, num_minor_ticks = 5, clear = FALSE)
}
\arguments{
\item{ticks}{List of tick locations.}

\item{num_minor_ticks}{The number of minor tick positions to generate between adjacent major tick values.}

\item{clear}{A logical indicating whether all non-specified parameters should be cleared out of any exising specification in the figure. This can be useful when modifying a previously-defined figure.}
}
\description{
Generate ticks at fixed, explicitly supplied locations.
}
\examples{
figure() \%>\%
  ly_points(1:10) \%>\%
  x_axis(ticker = ticker_fixed(c(1, 7, 9)))
}
