% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_axis_grid.R
\name{ticker_interval}
\alias{ticker_interval}
\title{Generate evenly spaced ticks at a fixed interval regardless of scale.}
\usage{
ticker_interval(interval = NULL, desired_num_ticks = 6,
  num_minor_ticks = 5, clear = FALSE)
}
\arguments{
\item{interval}{The interval between adjacent ticks.}

\item{desired_num_ticks}{A desired target number of major tick positions to generate across the plot range. Note: This value is a suggestion, and ticker subclasses may ignore it entirely, or use it only as an ideal goal to approach as well as can be, in the context of a specific ticking strategy.}

\item{num_minor_ticks}{The number of minor tick positions to generate between adjacent major tick values.}

\item{clear}{A logical indicating whether all non-specified parameters should be cleared out of any exising specification in the figure. This can be useful when modifying a previously-defined figure.}
}
\value{
A specification that is used as the \code{ticker} argument for \code{\link{x_axis}} or \code{\link{y_axis}}.
}
\description{
Generate evenly spaced ticks at a fixed interval regardless of scale.
}
\examples{
figure() \%>\%
  ly_points(1:10) \%>\%
  x_axis(ticker = ticker_interval(0.5))
}
