% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_axis_grid.R
\name{ticker_log}
\alias{ticker_log}
\title{Generate ticks on a log scale.}
\usage{
ticker_log(base = 10, min_interval = 0, max_interval = NULL,
  mantissas = list(1, 5), desired_num_ticks = 6, num_minor_ticks = 5,
  clear = FALSE)
}
\arguments{
\item{base}{The multiplier to use for scaling mantissas.}

\item{min_interval}{The smallest allowable interval between two adjacent ticks.}

\item{max_interval}{The largest allowable interval between two adjacent ticks. Note: To specify an unbounded interval, set to \code{NULL}.}

\item{mantissas}{The acceptable list numbers to generate multiples of.}

\item{desired_num_ticks}{A desired target number of major tick positions to generate across the plot range. Note: This value is a suggestion, and ticker subclasses may ignore it entirely, or use it only as an ideal goal to approach as well as can be, in the context of a specific ticking strategy.}

\item{num_minor_ticks}{The number of minor tick positions to generate between adjacent major tick values.}

\item{clear}{A logical indicating whether all non-specified parameters should be cleared out of any exising specification in the figure. This can be useful when modifying a previously-defined figure.}
}
\value{
A specification that is used as the \code{ticker} argument for \code{\link{x_axis}} or \code{\link{y_axis}}.
}
\description{
Generate ticks on a log scale.
}
\examples{
figure() \%>\%
  ly_points(rexp(100, rate = 1 / 100), rnorm(100)) \%>\%
  x_axis(ticker = ticker_log(base = 2))
}
