% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_axis_grid.R
\name{ticker_num}
\alias{ticker_num}
\title{Generate ticks on a linear / numeric scale.}
\usage{
ticker_num(min_interval = 0, max_interval = NULL, desired_num_ticks = 6,
  num_minor_ticks = 5, mantissas = c(1, 2, 5), base = 10, clear = FALSE)
}
\arguments{
\item{min_interval}{The smallest allowable interval between two adjacent ticks.}

\item{max_interval}{The largest allowable interval between two adjacent ticks. Note: To specify an unbounded interval, set to \code{NULL}.}

\item{desired_num_ticks}{A desired target number of major tick positions to generate across the plot range. Note: This value is a suggestion, and ticker subclasses may ignore it entirely, or use it only as an ideal goal to approach as well as can be, in the context of a specific ticking strategy.}

\item{num_minor_ticks}{The number of minor tick positions to generate between adjacent major tick values.}

\item{mantissas}{The acceptable list numbers to generate multiples of.}

\item{base}{The multiplier to use for scaling mantissas.}

\item{clear}{A logical indicating whether all non-specified parameters should be cleared out of any exising specification in the figure. This can be useful when modifying a previously-defined figure.}
}
\value{
A specification that is used as the \code{ticker} argument for \code{\link{x_axis}} or \code{\link{y_axis}}.
}
\description{
Generate ticks on a linear / numeric scale.
}
