% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_axis_grid.R
\name{tickformatter_log}
\alias{tickformatter_log}
\title{Display tick values from continuous ranges as powers of some base. Most often useful in conjunction with a "LogTicker".}
\usage{
tickformatter_log(clear = FALSE)
}
\arguments{
\item{clear}{A logical indicating whether all non-specified parameters should be cleared out of any exising specification in the figure. This can be useful when modifying a previously-defined figure.}
}
\value{
A specification that is used as the \code{tickformatter} argument for \code{\link{x_axis}} or \code{\link{y_axis}}.
}
\description{
Display tick values from continuous ranges as powers of some base. Most often useful in conjunction with a "LogTicker".
}
\note{
This function should never need to be used as there are no parameters for a log tickformatter.
}
