% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_axis_grid.R
\name{tickformatter_num}
\alias{tickformatter_num}
\title{Display tick values from continuous ranges as "basic numbers", using scientific notation when appropriate by default.}
\usage{
tickformatter_num(use_scientific = TRUE, power_limit_low = -3,
  power_limit_high = 5, precision = "auto", clear = FALSE)
}
\arguments{
\item{use_scientific}{Whether to ever display scientific notation. If \code{TRUE}, then when to use scientific notation is controlled by "power_limit_low" and "power_limit_high".}

\item{power_limit_low}{Limit the use of scientific notation to when: log(x) <= power_limit_low}

\item{power_limit_high}{Limit the use of scientific notation to when: log(x) >= power_limit_high}

\item{precision}{How many digits of precision to display in tick labels.}

\item{clear}{A logical indicating whether all non-specified parameters should be cleared out of any exising specification in the figure. This can be useful when modifying a previously-defined figure.}
}
\value{
A specification that is used as the \code{tickformatter} argument for \code{\link{x_axis}} or \code{\link{y_axis}}.
}
\description{
Display tick values from continuous ranges as "basic numbers", using scientific notation when appropriate by default.
}
\examples{
figure() \%>\%
  ly_points(c(1:10) * 1000000, 1:10) \%>\%
  x_axis(tickformatter = tickformatter_num(use_scientific = FALSE))
}
