% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_axis_grid.R
\name{tickformatter_numeral}
\alias{tickformatter_numeral}
\title{Tick formatter based on a human-readable format string.}
\usage{
tickformatter_numeral(format = "0,0", rounding = "round", language = "en",
  clear = FALSE)
}
\arguments{
\item{format}{The number format, as defined in the tables described in the details. For the complete specification, see \url{http://numbrojs.com/format.html}.}

\item{rounding}{Rounding functions (round, floor, ceil) and their synonyms (nearest, rounddown, roundup).}

\item{language}{The language to use for formatting language-specific features (e.g. thousands separator).}

\item{clear}{A logical indicating whether all non-specified parameters should be cleared out of any exising specification in the figure. This can be useful when modifying a previously-defined figure.}
}
\value{
A specification that is used as the \code{tickformatter} argument for \code{\link{x_axis}} or \code{\link{y_axis}}.
}
\description{
Tick formatter based on a human-readable format string.
}
\details{
Possible values for \code{format} are defined in the following tables:

\strong{NUMBERS}

\tabular{llr}{
\strong{Number}    \tab  \strong{Format}      \tab  \strong{String}\cr
\code{10000}       \tab  \code{'0,0.0000'}    \tab  \code{10,000.0000}\cr
\code{10000.23}    \tab  \code{'0,0'}         \tab  \code{10,000}\cr
\code{10000.23}    \tab  \code{'+0,0'}        \tab  \code{+10,000}\cr
\code{-10000}      \tab  \code{'0,0.0'}       \tab  \code{-10,000.0}\cr
\code{10000.1234}  \tab  \code{'0.000'}       \tab  \code{10000.123}\cr
\code{10000.1234}  \tab  \code{'0[.]00000'}   \tab  \code{10000.12340}\cr
\code{-10000}      \tab  \code{'(0,0.0000)'}  \tab  \code{(10,000.0000)}\cr
\code{-0.23}       \tab  \code{'.00'}         \tab  \code{-.23}\cr
\code{-0.23}       \tab  \code{'(.00)'}       \tab  \code{(.23)}\cr
\code{0.23}        \tab  \code{'0.00000'}     \tab  \code{0.23000}\cr
\code{0.23}        \tab  \code{'0.0[0000]'}   \tab  \code{0.23}\cr
\code{1230974}     \tab  \code{'0.0a'}        \tab  \code{1.2m}\cr
\code{1460}        \tab  \code{'0 a'}         \tab  \code{1 k}\cr
\code{-104000}     \tab  \code{'0a'}          \tab  \code{-104k}\cr
\code{1}           \tab  \code{'0o'}          \tab  \code{1st}\cr
\code{52}          \tab  \code{'0o'}          \tab  \code{52nd}\cr
\code{23}          \tab  \code{'0o'}          \tab  \code{23rd}\cr
\code{100}         \tab  \code{'0o'}          \tab  \code{100th}
}

\strong{CURRENCY}

\tabular{llr}{
\strong{Number}   \tab  \strong{Format}      \tab   \strong{String}\cr
\code{1000.234}   \tab  \code{'$0,0.00'}     \tab  \code{ $1,000.23}\cr
\code{1000.2}     \tab  \code{'0,0[.]00 $'}  \tab  \code{ 1,000.20 $}\cr
\code{1001}       \tab  \code{'$ 0,0[.]00'}  \tab  \code{ $ 1,001}\cr
\code{-1000.234}  \tab  \code{'($0,0)'}      \tab  \code{ ($1,000)}\cr
\code{-1000.234}  \tab  \code{'$0.00'}       \tab  \code{ -$1000.23}\cr
\code{1230974}    \tab  \code{'($ 0.00 a)'}  \tab  \code{ $ 1.23 m}
}

\strong{BYTES}

\tabular{llr}{
\strong{Number}       \tab \strong{Format}   \tab \strong{String}\cr
\code{100}            \tab \code{'0b'}       \tab \code{100B}\cr
\code{2048}           \tab \code{'0 b'}      \tab \code{2 KB}\cr
\code{7884486213}     \tab \code{'0.0b'}     \tab \code{7.3GB}\cr
\code{3467479682787}  \tab \code{'0.000 b'}  \tab \code{3.154 TB}
}

\strong{PERCENTAGES}

\tabular{llr}{
\strong{Number}     \tab  \strong{Format}      \tab  \strong{String}\cr
\code{1}            \tab  \code{'0\%'}         \tab  \code{100\%}\cr
\code{0.974878234}  \tab  \code{'0.000\%'}     \tab  \code{97.488\%}\cr
\code{-0.43}        \tab  \code{'0 \%'}        \tab  \code{-43 \%}\cr
\code{0.43}         \tab  \code{'(0.000 \%)'}  \tab  \code{43.000 \%}
}

\strong{TIME}

\tabular{llr}{
\strong{Number}  \tab \strong{Format}    \tab   \strong{String}\cr
\code{25}        \tab \code{'00:00:00'}  \tab   \code{0:00:25}\cr
\code{238}       \tab \code{'00:00:00'}  \tab   \code{0:03:58}\cr
\code{63846}     \tab \code{'00:00:00'}  \tab   \code{17:44:06}
}
}
\examples{
figure() \%>\%
  ly_points(1:10) \%>\%
  x_axis(tickformatter = tickformatter_numeral(format = "0o"))

figure() \%>\%
  ly_points((1:10) ^ 4, 1:10) \%>\%
  x_axis(tickformatter = tickformatter_numeral(format = "($0,0)", language = "it"))

figure() \%>\%
  ly_points((1:11) ^ 6, 1:11) \%>\%
  x_axis(tickformatter = tickformatter_numeral(format = "0.0b"))
}
