% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_axis_grid.R
\name{tickformatter_printf}
\alias{tickformatter_printf}
\title{Tick formatter based on a printf-style format string.}
\usage{
tickformatter_printf(format = "\%s", clear = FALSE)
}
\arguments{
\item{format}{The number format, as defined in the details below.}

\item{clear}{A logical indicating whether all non-specified parameters should be cleared out of any exising specification in the figure. This can be useful when modifying a previously-defined figure.}
}
\value{
A specification that is used as the \code{tickformatter} argument for \code{\link{x_axis}} or \code{\link{y_axis}}.
}
\description{
Tick formatter based on a printf-style format string.
}
\details{
The placeholder in the format string is marked by % and is followed by one or more of these elements, in this order:
\itemize{
  \item An optional "+" sign Causes the result to be preceded with a plus or minus sign on numeric values. By default, only the "-" sign is used on negative numbers.
  \item An optional padding specifier Specifies what (if any) character to use for padding. Possible values are 0 or any other character preceded by a "'" (single quote). The default is to pad with spaces.
  \item An optional "-" sign Causes sprintf to left-align the result of this placeholder. The default is to right-align the result.
  \item An optional number Specifies how many characters the result should have. If the value to be returned is shorter than this number, the result will be padded.
  \item An optional precision modifier Consists of a "." (dot) followed by a number, specifies how many digits should be displayed for floating point numbers. When used on a string, it causes the result to be truncated.
  \item A type specifier Can be any of:
  \itemize{
    \item "\%" --- yields a literal "\%" character
    \item "b" --- yields an integer as a binary number
    \item "c" --- yields an integer as the character with that ASCII value
    \item "d" or "i" --- yields an integer as a signed decimal number
    \item "e" --- yields a float using scientific notation
    \item "u" --- yields an integer as an unsigned decimal number
    \item "f" --- yields a float as is
    \item "o" --- yields an integer as an octal number
    \item "s" --- yields a string as is
    \item "x" --- yields an integer as a hexadecimal number (lower-case)
    \item "X" --- yields an integer as a hexadecimal number (upper-case
  }
}
}
\examples{
figure() \%>\%
  ly_points(1:10) \%>\%
  x_axis(tickformatter = tickformatter_printf(format = "\%5.3f mu"))
}
