% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_tiles.R
\name{tile_spec}
\alias{tile_spec}
\title{Specify a tile provider for a tile map plot.}
\usage{
tile_spec(url, attribution, type = "WMTS")
}
\arguments{
\item{url}{Tile service url e.g., http://c.tile.openstreetmap.org/{Z}/{X}/{Y}.png}

\item{attribution}{Data provider attribution content. This can include HTML content.}

\item{type}{The type of tile provider (currently only "WMTS" is supported).}
}
\description{
Specify a tile provider for a tile map plot.
}
