% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{to_epoch}
\alias{to_epoch}
\title{Convert a date or POSIXct to Unix epoch time.}
\usage{
to_epoch(x)
}
\arguments{
\item{x}{Date or POSIXct vector.}
}
\description{
Convert a date or POSIXct to Unix epoch time.
}
\note{
Date-type data is automatically converted (since Bokeh expects dates in epoch time) using this function when a plot is being built, so this function should rarely need to be called explicitly.
}
