% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_tools.R
\name{tool_box_zoom}
\alias{tool_box_zoom}
\title{Add "box_zoom" tool to a Bokeh figure}
\usage{
tool_box_zoom(fig, dimensions = NULL, match_aspect = FALSE,
  line_color = NULL, line_alpha = NULL, fill_color = NULL,
  fill_alpha = NULL, line_width = NULL, line_dash = NULL,
  line_dash_offset = NULL, line_join = NULL, line_cap = NULL,
  remove = FALSE)
}
\arguments{
\item{dimensions}{Which dimensions the zoom box is to be free in. By default, users may freely draw zoom boxes with any dimensions. If only "width" is supplied, the box will be constrained to span the entire vertical space of the plot, only the horizontal dimension can be controlled. If only "height" is supplied, the box will be constrained to span the entire horizontal space of the plot, and the vertical dimension can be controlled. One of 'width', 'height', or 'both'.}

\item{match_aspect}{Whether the box zoom region should be restricted to have the same aspect ratio as the plot region. Boolean. Note: If the tool is restricted to one dimension, this value has no effect.}
}
\description{
Add "box_zoom" tool to a Bokeh figure
}
