% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_tools.R
\name{tool_crosshair}
\alias{tool_crosshair}
\title{Add "crosshair" tool to a Bokeh figure}
\usage{
tool_crosshair(fig, dimensions = NULL, line_width = NULL,
  line_alpha = NULL, line_color = NULL, remove = FALSE)
}
\arguments{
\item{fig}{Figure to modify.}

\item{dimensions}{Which dimensions the crosshair tool is to track. By default, both a vertical and horizontal line will be drawn. If only "width" is supplied, only a horizontal line will be drawn. If only "height" is supplied, only a vertical line will be drawn. One of 'width', 'height', or 'both'.}

\item{line_width}{Stroke width in units of pixels.}

\item{line_alpha}{An alpha value to use to stroke paths with. Acceptable values are floating point numbers between 0 (transparent) and 1 (opaque).}

\item{line_color}{A color to use to stroke paths with.}

\item{remove}{Logical indicating whether the tool should be removed.}
}
\description{
Add "crosshair" tool to a Bokeh figure
}
\examples{
figure(tools = NULL) \%>\%
  ly_points(1:10) \%>\%
  tool_crosshair(line_width = 12, line_alpha = 0.25)
}
\seealso{
Other tools: \code{\link{tool_box_select}},
  \code{\link{tool_box_zoom}}, \code{\link{tool_help}},
  \code{\link{tool_lasso_select}}, \code{\link{tool_pan}},
  \code{\link{tool_poly_select}}, \code{\link{tool_redo}},
  \code{\link{tool_reset}}, \code{\link{tool_save}},
  \code{\link{tool_undo}}, \code{\link{tool_wheel_pan}},
  \code{\link{tool_wheel_zoom}},
  \code{\link{tool_zoom_in}}, \code{\link{tool_zoom_out}}
}
