% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_tools.R
\name{tool_lasso_select}
\alias{tool_lasso_select}
\title{Add "lasso_select" tool to a Bokeh figure}
\usage{
tool_lasso_select(fig, select_every_mousemove = TRUE, line_color = NULL,
  line_alpha = NULL, fill_color = NULL, fill_alpha = NULL,
  line_width = NULL, line_dash = NULL, line_dash_offset = NULL,
  line_join = NULL, line_cap = NULL, names = NULL, callback = NULL,
  remove = FALSE)
}
\arguments{
\item{select_every_mousemove}{Whether a selection computation should happen on every mouse event, or only once, when the selection region is completed. Default: TRUE.}

\item{names}{A list of names to query for. If set, only renderers that have a matching value for their ``name`` attribute will be used.}

\item{callback}{A callback to run in the browser on every selection of a lasso area. The cb_data parameter that is available to the Callback code will contain one LassoSelectTool-specific field.}
}
\description{
Add "lasso_select" tool to a Bokeh figure
}
