% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_tools.R
\name{tool_pan}
\alias{tool_pan}
\title{Add "pan" tool to a Bokeh figure}
\usage{
tool_pan(fig, dimensions = NULL, remove = FALSE)
}
\arguments{
\item{fig}{Figure to modify.}

\item{dimensions}{Which dimensions the pan tool is constrained to act in. By default the pan tool will pan in any dimension, but can be configured to only pan horizontally across the width of the plot, or vertically across the height of the plot. One of 'width', 'height', or 'both'.}

\item{remove}{Logical indicating whether the tool should be removed.}
}
\description{
Add "pan" tool to a Bokeh figure
}
\examples{
\donttest{
# only pan on x axis
figure() \%>\% ly_points(1:10) \%>\%
 tool_pan(dimensions = "height")
}
figure(tools = NULL) \%>\%
  ly_points(1:10) \%>\%
  tool_pan(dimensions = "width")
}
\seealso{
Other tools: \code{\link{tool_box_select}},
  \code{\link{tool_box_zoom}},
  \code{\link{tool_crosshair}}, \code{\link{tool_help}},
  \code{\link{tool_lasso_select}},
  \code{\link{tool_poly_select}}, \code{\link{tool_redo}},
  \code{\link{tool_reset}}, \code{\link{tool_save}},
  \code{\link{tool_undo}}, \code{\link{tool_wheel_pan}},
  \code{\link{tool_wheel_zoom}},
  \code{\link{tool_zoom_in}}, \code{\link{tool_zoom_out}}
}
