% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_tools.R
\name{tool_reset}
\alias{tool_reset}
\title{Add "reset" tool to a Bokeh figure}
\usage{
tool_reset(fig, reset_size = TRUE, remove = FALSE)
}
\arguments{
\item{reset_size}{Whether activating the Reset tool should also reset the plot's canvas dimensions to their original size.}
}
\description{
Add "reset" tool to a Bokeh figure
}
