% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_selection.R
\name{tool_selection}
\alias{tool_selection}
\title{Add "selection" tool callback to a Bokeh figure}
\usage{
tool_selection(fig, callback, ref_layer)
}
\arguments{
\item{fig}{figure to modify}

\item{callback}{a callback to be applied to this tool - either a character string of javascript code or any one of \code{\link{debug_callback}}, \code{\link{shiny_callback}}, \code{\link{console_callback}}, \code{\link{custom_callback}}}

\item{ref_layer}{name of the layer that the callback should be applied to}
}
\description{
This adds a selection callback to be used with the box select or lasso select tools.
}
