% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_tools.R
\name{tool_wheel_pan}
\alias{tool_wheel_pan}
\title{Add "wheel_pan" tool to a Bokeh figure}
\usage{
tool_wheel_pan(fig, dimension = NULL, remove = FALSE)
}
\arguments{
\item{Which}{dimension the wheel pan tool is constrained to act in. By default the wheel pan tool will pan the plot along the x-axis. Must be one of 'width' or 'height'.}
}
\description{
Add "wheel_pan" tool to a Bokeh figure
}
