% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_tools.R
\name{tool_wheel_zoom}
\alias{tool_wheel_zoom}
\title{Add "wheel_zoom" tool to a Bokeh figure}
\usage{
tool_wheel_zoom(fig, dimensions = NULL, remove = FALSE)
}
\arguments{
\item{dimensions}{Which dimensions the wheel zoom tool is constrained to act in. By default the wheel zoom tool will zoom in any dimension, but can be configured to only zoom horizontally across the width of the plot, or vertically across the height of the plot. One of 'width', 'height', or 'both'.}
}
\description{
Add "wheel_zoom" tool to a Bokeh figure
}
