% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_tools.R
\name{tool_zoom_in}
\alias{tool_zoom_in}
\title{Add "zoom_in" tool to a Bokeh figure}
\usage{
tool_zoom_in(fig, dimensions = NULL, factor = NULL, remove = FALSE)
}
\arguments{
\item{dimensions}{Which dimensions the zoom-in tool is constrained to act in. By default the zoom-in tool will zoom in any dimension, but can be configured to only zoom horizontally across the width of the plot, or vertically across the height of the plot. One of 'width', 'height', or 'both'.}

\item{factor}{Proportion to zoom for each click of the zoom-in tool.}
}
\description{
Add "zoom_in" tool to a Bokeh figure
}
