% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_tools.R
\name{tool_zoom_out}
\alias{tool_zoom_out}
\title{Add "zoom_out" tool to a Bokeh figure}
\usage{
tool_zoom_out(fig, dimensions = NULL, factor = NULL, remove = FALSE)
}
\arguments{
\item{fig}{Figure to modify.}

\item{dimensions}{Which dimensions the zoom-out tool is constrained to act in. By default the zoom-out tool will zoom in any dimension, but can be configured to only zoom horizontally across the width of the plot, or vertically across the height of the plot. One of 'width', 'height', or 'both'.}

\item{factor}{Proportion to zoom for each click of the zoom-out tool.}

\item{remove}{Logical indicating whether the tool should be removed.}
}
\description{
Add "zoom_out" tool to a Bokeh figure
}
\examples{
figure(tools = NULL) \%>\%
  ly_points(1:10) \%>\%
  tool_zoom_in(factor = 0.5) \%>\%
  tool_zoom_out(factor = 0.5)
}
