% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fig_resize.R
\name{update_size}
\alias{update_size}
\title{Update the dimensions of a bokeh plot}
\usage{
update_size(fig, width, height)
}
\arguments{
\item{fig}{figure to resize}

\item{width}{desired width of figure, in pixels}

\item{height}{desired height of figure, in pixels}
}
\description{
Update the dimensions of a bokeh plot
}

