% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_axis_grid.R
\name{x_axis}
\alias{x_axis}
\title{Customize x axis of a Bokeh figure}
\usage{
x_axis(fig, label = NULL, position = "below", log = FALSE, draw = TRUE,
  ticker = NULL, tickformatter = NULL, axis = NULL, grid = NULL)
}
\description{
Customize x axis of a Bokeh figure
}
