% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_tiles.R
\name{add_tiles}
\alias{add_tiles}
\title{Add map tiles to a Bokeh figure.}
\usage{
add_tiles(fig, tile_spec, hi_res = TRUE, y_origin_offset = NULL,
  x_origin_offset = NULL, min_zoom = NULL, max_zoom = NULL,
  extra_url_vars = NULL, initial_resolution = NULL, tile_size = NULL,
  alpha = NULL, render_parents = NULL, visible = NULL, level = NULL)
}
\arguments{
\item{fig}{figure to modify}

\item{tile_spec}{A character string indicating to use one of the pre-defined tile specifications (see details) or a custom tile specification from calling \code{\link{tile_spec}}.}

\item{hi_res}{When using a pre-defined tile specification, should a high-resolution (if available) version of the tiles be served? Boolean.}

\item{y_origin_offset}{A y-offset in plot coordinates}

\item{x_origin_offset}{An x-offset in plot coordinates}

\item{min_zoom}{A minimum zoom level for the tile layer. This is the most zoomed-out level.}

\item{max_zoom}{A maximum zoom level for the tile layer. This is the most zoomed-in level.}

\item{extra_url_vars}{A dictionary that maps url variable template keys to values. These variables are useful for parts of tile urls which do not change from tile to tile (e.g. server host name, or layer name).}

\item{initial_resolution}{Resolution (plot_units / pixels) of minimum zoom level of tileset projection. NULL to auto-compute.}

\item{tile_size}{Tile size in pixels (e.g. 256)}

\item{alpha}{Tile opacity 0.0 - 1.0.}

\item{render_parents}{Flag enable/disable drawing of parent tiles while waiting for new tiles to arrive. Default value is TRUE.}

\item{visible}{Is the renderer visible.}

\item{level}{Specifies the level in which to paint this renderer. One of 'image', 'underlay', 'glyph', 'annotation', 'overlay'.}
}
\description{
Add map tiles to a Bokeh figure.
}
