% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_annotation.R
\name{ann_box}
\alias{ann_box}
\title{Add a "box" annotation to a Bokeh figure}
\usage{
ann_box(fig, left = NULL, bottom = NULL, right = NULL, top = NULL,
  left_units = NULL, bottom_units = NULL, right_units = NULL,
  top_units = NULL, color = NULL, alpha = NULL, lgroup = NULL,
  lname = NULL, ...)
}
\arguments{
\item{fig}{Figure to modify.}
}
\description{
Add a "box" annotation to a Bokeh figure
}
