% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_annotation.R
\name{arrow}
\alias{arrow}
\title{Specify details of an arrow head.}
\usage{
arrow(type = c("normal", "open", "tee", "vee"), size = NULL, color = NULL,
  alpha = NULL, line_width = NULL, line_join = NULL, line_cap = NULL,
  line_dash = NULL, line_dash_offset = NULL, ...)
}
\arguments{
\item{fig}{Figure to modify.}
}
\description{
Specify details of an arrow head.
}
\seealso{
Other annotation functions: \code{\link{ann_arrow}},
  \code{\link{ann_band}}, \code{\link{ann_box}},
  \code{\link{ann_labels}}, \code{\link{ann_poly}},
  \code{\link{ann_span}}, \code{\link{ann_title}},
  \code{\link{ann_whisker}}
}
