% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure.R
\name{figure}
\alias{figure}
\title{Instantiate an rbokeh figure}
\usage{
figure(data = NULL, width = 600, height = 600, title = NULL,
  xlab = NULL, ylab = NULL, xlim = NULL, ylim = NULL,
  padding_factor = 0.07, xgrid = TRUE, ygrid = TRUE,
  legend_location = "top_right", tools = c("pan", "wheel_zoom", "box_zoom",
  "reset", "save", "help"), logo = NULL, theme = getOption("bokeh_theme"),
  toolbar_location = "above", h_symmetry = TRUE, v_symmetry = FALSE,
  lod_factor = 10, lod_interval = 300, lod_threshold = NULL,
  lod_timeout = 500)
}
\arguments{
\item{data}{default dataset to use}
}
\description{
Instantiate an rbokeh figure
}
