% Generated by roxygen2 (4.0.2): do not edit by hand
\name{figure}
\alias{figure}
\title{Start a Bokeh Figure}
\usage{
figure(width = 480, height = 480, title = character(0),
  xlim = numeric(0), ylim = numeric(0), plot_width = numeric(0),
  plot_height = numeric(0), x_axis_type = character(0),
  y_axis_type = character(0), x_mapper_type = character(0),
  y_mapper_type = character(0), background_fill = character(0),
  border_fill = character(0), min_border = numeric(0),
  min_border_left = numeric(0), min_border_right = numeric(0),
  min_border_top = numeric(0), min_border_bottom = numeric(0),
  h_symmetry = logical(0), v_symmetry = logical(0),
  outline_line_color = character(0), xaxes = "below", yaxes = "left",
  tools = character(0), theme = getOption("bokeh_theme"))
}
\arguments{
\item{width}{figure width in pixels}

\item{height}{figure width in pixels}

\item{title}{a title to display above the plot. - "title" is also the prefix for a set of Text Properties, so you can set the font for the title with the parameter text_font.}

\item{xlim}{the extent of the plotting area in the x-dimension (will be computed automatically if not specified).}

\item{ylim}{the extent of the plotting area in the y-dimension (will be computed automatically if not specified).}

\item{plot_width,plot_height}{width and height of the entire plot in pixels, including border space}

\item{x_axis_type,y_axis_type}{can be set to "datetime" to create datetime axis}

\item{x_mapper_type,y_mapper_type}{can be set to "log" to specifically set the mapper used for the axis}

\item{background_fill}{a color to fill the inner plot area with}

\item{border_fill}{a color to fill the border region around the plot area with.}

\item{min_border}{a minimum size in pixels for the border. This applies to all sides of the plot.}

\item{min_border_left}{set left border individually}

\item{min_border_right}{set right border individually}

\item{min_border_top}{set top border individually}

\item{min_border_bottom}{set bottom border individually}

\item{h_symmetry,v_symmetry}{whether to symmetrize plot borders on opposite horizontal or vertical sides of the plot.}

\item{outline_line_color}{Line Properties that controls the appearance of an outline around the plot, for instance you can set the color of the outline with this}

\item{xaxes}{where to put x axis labels}

\item{yaxes}{where to put y axis labels}

\item{tools}{interactivity tools options}

\item{theme}{an rbokeh theme to use (tableau by default)}
}
\description{
Start a Bokeh Figure
}

