% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/figure.R
\name{figure}
\alias{figure}
\title{Initialize a Bokeh figure}
\usage{
figure(width = 480, height = 480, title = NULL, xlab = NULL,
  ylab = NULL, xlim = NULL, ylim = NULL, padding_factor = 0.07,
  plot_width = NULL, plot_height = NULL, xgrid = TRUE, ygrid = TRUE,
  xaxes = "below", yaxes = "left", tools = c("pan", "wheel_zoom",
  "box_zoom", "resize", "reset", "save"), theme = getOption("bokeh_theme"),
  ...)
}
\arguments{
\item{width}{figure width in pixels}

\item{height}{figure width in pixels}

\item{title}{a title to display above the plot. - "title" is also the prefix for a set of Text Properties, so you can set the font for the title with the parameter text_font.}

\item{xlab}{label for x axis}

\item{ylab}{label for y axis}

\item{xlim}{the extent of the plotting area in the x-dimension (will be computed automatically if not specified).}

\item{ylim}{the extent of the plotting area in the y-dimension (will be computed automatically if not specified).}

\item{padding_factor}{if limits are not specified, by what factor should the extents of the data be padded}

\item{plot_width,plot_height}{width and height of the entire plot in pixels, including border space}

\item{xgrid}{whether to draw x axis grid lines}

\item{ygrid}{whether to draw y axis grid lines}

\item{xaxes}{where to put x axis, or FALSE if no x axis ticks / labels}

\item{yaxes}{where to put y axis, or FALSE if no y axis ticks / labels}

\item{tools}{interactivity tools options}

\item{theme}{an rbokeh theme to use (tableau by default)}

\item{\ldots}{additional parameters for fine control over axes (see "Additional parameters" below)}
}
\description{
Initialize a Bokeh figure
}
\section{Additional parameters}{

\tabular{ll}{
\code{axis_label_standoff} \tab (integer) The distance in pixels that the axis labels should be offset from the tick labels. \cr
\code{axis_label_text_align} \tab ('left', 'right', 'center') The text align of the axis label. \cr
\code{axis_label_text_alpha} \tab (numeric) The text alpha of the axis label. \cr
\code{axis_label_text_baseline} \tab ('top', 'middle', 'bottom', 'alphabetic', 'hanging') The text baseline of the axis label. \cr
\code{axis_label_text_color} \tab (color) The text color of the axis label. \cr
\code{axis_label_text_font} \tab (string) The text font of the axis label. \cr
\code{axis_label_text_font_size} \tab (string - e.g. '12pt') The text font size of the axis label. \cr
\code{axis_label_text_font_style} \tab ('normal', 'italic', 'bold') The text font style of the axis label. \cr
\code{axis_line_alpha} \tab (numeric) The line alpha of the axis line. \cr
\code{axis_line_cap} \tab ('butt', 'round', 'square') The line cap of the axis line. \cr
\code{axis_line_color} \tab (color) The line color of the axis line. \cr
\code{axis_line_dash} \tab The line dash of the axis line. \cr
\code{axis_line_dash_offset} \tab (integer) The line dash offset of the axis line. \cr
\code{axis_line_join} \tab ('miter', 'round', 'bevel') The line join of the axis line. \cr
\code{axis_line_width} \tab (integer) The line width of the axis line. \cr
\code{major_label_orientation} \tab ('horizontal', 'vertical', or angle in degrees) What direction the major label text should be oriented. If a number is supplied, the angle of the text is measured from horizontal. \cr
\code{major_label_standoff} \tab (integer) The distance in pixels that the major tick labels should be offset from the associated ticks. \cr
\code{major_label_text_align} \tab ('left', 'right', 'center') The text align of the major tick labels. \cr
\code{major_label_text_alpha} \tab (numeric) The text alpha of the major tick labels. \cr
\code{major_label_text_baseline} \tab ('top', 'middle', 'bottom', 'alphabetic', 'hanging') The text baseline of the major tick labels. \cr
\code{major_label_text_color} \tab (color) The text color of the major tick labels. \cr
\code{major_label_text_font} \tab (string - 'Helvetica') The text font of the major tick labels. \cr
\code{major_label_text_font_size} \tab (string - e.g. '12pt') The text font size of the major tick labels. \cr
\code{major_label_text_font_style} \tab ('normal', 'italic', 'bold') The text font style of the major tick labels. \cr
\code{major_tick_in} \tab (integer) The distance in pixels that major ticks should extend into the main plot area. \cr
\code{major_tick_line_alpha} \tab (numeric) The line alpha of the major ticks. \cr
\code{major_tick_line_cap} \tab ('butt', 'round', 'square') The line cap of the major ticks. \cr
\code{major_tick_line_color} \tab (color) The line color of the major ticks. \cr
\code{major_tick_line_dash} \tab The line dash of the major ticks. \cr
\code{major_tick_line_dash_offset} \tab (integer) The line dash offset of the major ticks. \cr
\code{major_tick_line_join} \tab ('miter', 'round', 'bevel') The line join of the major ticks. \cr
\code{major_tick_line_width} \tab (integer) The line width of the major ticks. \cr
\code{major_tick_out} \tab (integer) The distance in pixels that major ticks should extend out of the main plot area. \cr
\code{minor_tick_in} \tab (integer) The distance in pixels that minor ticks should extend into the main plot area. \cr
\code{minor_tick_line_alpha} \tab (numeric) The line alpha of the minor ticks. \cr
\code{minor_tick_line_cap} \tab ('butt', 'round', 'square') The line cap of the minor ticks. \cr
\code{minor_tick_line_color} \tab (color) The line color of the minor ticks. \cr
\code{minor_tick_line_dash} \tab The line dash of the minor ticks. \cr
\code{minor_tick_line_dash_offset} \tab (integer) The line dash offset of the minor ticks. \cr
\code{minor_tick_line_join} \tab ('miter', 'round', 'bevel') The line join of the minor ticks. \cr
\code{minor_tick_line_width} \tab (integer) The line width of the minor ticks. \cr
\code{minor_tick_out} \tab (integer) The distance in pixels that major ticks should extend out of the main plot area. \cr
 }
}

