% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/figure.R
\name{figure}
\alias{figure}
\title{Initialize a Bokeh figure}
\usage{
figure(width = 480, height = 480, title = NULL, xlab = NULL,
  ylab = NULL, xlim = NULL, ylim = NULL, padding_factor = 0.07,
  plot_width = NULL, plot_height = NULL, xgrid = TRUE, ygrid = TRUE,
  xaxes = "below", yaxes = "left", tools = c("pan", "wheel_zoom",
  "box_zoom", "resize", "reset", "save"), theme = getOption("bokeh_theme"),
  ...)
}
\arguments{
\item{width}{figure width in pixels}

\item{height}{figure width in pixels}

\item{title}{a title to display above the plot. - "title" is also the prefix for a set of Text Properties, so you can set the font for the title with the parameter text_font.}

\item{xlab}{label for x axis}

\item{ylab}{label for y axis}

\item{xlim}{the extent of the plotting area in the x-dimension (will be computed automatically if not specified).}

\item{ylim}{the extent of the plotting area in the y-dimension (will be computed automatically if not specified).}

\item{padding_factor}{if limits are not specified, by what factor should the extents of the data be padded}

\item{plot_width,plot_height}{width and height of the entire plot in pixels, including border space}

\item{xgrid}{whether to draw x axis grid lines}

\item{ygrid}{whether to draw y axis grid lines}

\item{xaxes}{where to put x axis, or FALSE if no x axis ticks / labels}

\item{yaxes}{where to put y axis, or FALSE if no y axis ticks / labels}

\item{tools}{character vector of interactivity tools options (acceptable values are: "pan", "wheel_zoom", "box_zoom", "resize", "crosshair", "box_select", "lasso_select", "reset", "save").  Additionally, tool functions can be called on a figure to specify more control - see the "See Also" section below for a list of tool functions.  If \code{NULL}, the toolbar will not be drawn.  If \code{""} the toolbar will be drawn but no tools will be added by default.}

\item{theme}{an rbokeh theme to use (tableau by default)}

\item{\ldots}{additional parameters for fine control figure properties (see "Additional parameters" below)}
}
\description{
Initialize a Bokeh figure
}
\section{Additional parameters}{

\tabular{ll}{
\code{background_fill} \tab (color) background color of plot \cr
\code{border_fill} \tab (color) fill color of border area of plot \cr
\code{min_border} \tab (integer) A convenience property to set all all the min_X_border properties to the same value. If an individual border property is explicitly set, it will override min_border. \cr
\code{min_border_bottom} \tab (integer) Minimum size in pixels of the padding region below the bottom of the central plot region. This is a minimum. The padding region may expand as needed to accommodate titles or axes, etc. \cr
\code{min_border_left} \tab (integer) Minimum size in pixels of the padding region to the left of the central plot region. This is a minimum. The padding region may expand as needed to accommodate titles or axes, etc. \cr
\code{min_border_right} \tab (integer) Minimum size in pixels of the padding region to the right of the central plot region. This is a minimum. The padding region may expand as needed to accommodate titles or axes, etc. \cr
\code{min_border_top} \tab (integer) Minimum size in pixels of the padding region above the top of the central plot region. This is a minimum. The padding region may expand as needed to accommodate titles or axes, etc. \cr
\code{outline_line_alpha} \tab (numeric) The line alpha for the plot border outline. \cr
\code{outline_line_cap} \tab ('butt', 'round', 'square') The line cap for the plot border outline. \cr
\code{outline_line_color} \tab (color) The line color for the plot border outline. \cr
\code{outline_line_dash} \tab The line dash for the plot border outline. \cr
\code{outline_line_dash_offset} \tab (integer) The line dash offset for the plot border outline. \cr
\code{outline_line_join} \tab ('miter', 'round', 'bevel') The line join for the plot border outline. \cr
\code{outline_line_width} \tab (integer) The line width for the plot border outline. \cr
\code{plot_height} \tab (integer) Total height of the entire plot (including any axes, titles, border padding, etc.) This corresponds directly to the height of the HTML canvas that will be used. \cr
\code{plot_width} \tab (integer) Total width of the entire plot (including any axes, titles, border padding, etc.) This corresponds directly to the width of the HTML canvas that will be used. \cr
\code{title_text_align} \tab ('left', 'right', 'center') The text align for the plot title. \cr
\code{title_text_alpha} \tab The text alpha for the plot title. \cr
\code{title_text_baseline} \tab ('top', 'middle', 'bottom', 'alphabetic', 'hanging') The text baseline for the plot title. \cr
\code{title_text_color} \tab (color) The text color for the plot title. \cr
\code{title_text_font} \tab (string) The text font for the plot title. \cr
\code{title_text_font_size} \tab (string - e.g. '12pt') The text font size for the plot title. \cr
\code{title_text_font_style} \tab ('normal', 'italic', 'bold') The text font style for the plot title. \cr
\code{toolbar_location} \tab ('above', 'below', 'left', 'right') Where the toolbar will be located. If set to None, no toolbar will be attached to the plot. \cr
\code{h_symmetry} \tab (logical) Whether the total horizontal padding on both sides of the plot will be made equal (the left or right padding amount, whichever is larger). \cr
\code{v_symmetry} \tab (logical) Whether the total vertical padding on both sides of the plot will be made equal (the top or bottom padding amount, whichever is larger). \cr
\code{logo} \tab ('normal', 'grey') What version of the Bokeh logo to display on the toolbar. If set to None, no logo will be displayed. \cr
 }
}
\examples{
figure() \%>\% ly_points(1:10)
}
\seealso{
Layers to add to a figure: \code{\link{ly_abline}};
  \code{\link{ly_annular_wedge}}; \code{\link{ly_annulus}};
  \code{\link{ly_arc}}; \code{\link{ly_bezier}};
  \code{\link{ly_boxplot}}; \code{\link{ly_contour}};
  \code{\link{ly_crect}}; \code{\link{ly_curve}};
  \code{\link{ly_density}}; \code{\link{ly_hist}};
  \code{\link{ly_image_url}}; \code{\link{ly_image}};
  \code{\link{ly_lines}}; \code{\link{ly_map}};
  \code{\link{ly_multi_line}}; \code{\link{ly_oval}};
  \code{\link{ly_patch}}; \code{\link{ly_points}};
  \code{\link{ly_polygons}}; \code{\link{ly_quadratic}};
  \code{\link{ly_quantile}}; \code{\link{ly_ray}};
  \code{\link{ly_segments}}; \code{\link{ly_text}};
  \code{\link{ly_wedge}}
Tools to add to a figure: \code{\link{tool_box_select}};
  \code{\link{tool_box_zoom}};
  \code{\link{tool_crosshair}};
  \code{\link{tool_lasso_select}};
  \code{\link{tool_reset}}; \code{\link{tool_resize}};
  \code{\link{tool_save}}; \code{\link{tool_wheel_zoom}}
Other figure types: \code{\link{grid_plot}}; \code{\link{gmap}}
}

