% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fig_legend.R
\name{legend_control}
\alias{legend_control}
\title{Finer control over legend}
\usage{
legend_control(fig, orientation = "top_right", ...)
}
\arguments{
\item{fig}{figure to modify}

\item{orientation}{('top_right', 'top_left', 'bottom_left', 'bottom_right') The location where the legend should draw itself.}

\item{\ldots}{additional parameters for fine control of the legend (see "Additional parameters" below)}
}
\description{
Finer control over legend
}
\section{Additional parameters}{

\tabular{ll}{
\code{border_line_alpha} \tab The line alpha for the legend border outline. \cr
\code{border_line_cap} \tab ('butt', 'round', 'square') The line cap for the legend border outline. \cr
\code{border_line_color} \tab The line color for the legend border outline. \cr
\code{border_line_dash} \tab The line dash for the legend border outline. \cr
\code{border_line_dash_offset} \tab The line dash offset for the legend border outline. \cr
\code{border_line_join} \tab ('miter', 'round', 'bevel') The line join for the legend border outline. \cr
\code{border_line_width} \tab The line width for the legend border outline. \cr
\code{glyph_height} \tab The height (in pixels) that the rendered legend glyph should occupy. \cr
\code{glyph_width} \tab The width (in pixels) that the rendered legend glyph should occupy. \cr
\code{label_height} \tab The height (in pixels) of the area that legend labels should occupy. \cr
\code{label_standoff} \tab The distance (in pixels) to separate the label from its associated glyph. \cr
\code{label_text_align} \tab ('left', 'right', 'center') The text align for the legend labels. \cr
\code{label_text_alpha} \tab The text alpha for the legend labels. \cr
\code{label_text_baseline} \tab ('top', 'middle', 'bottom', 'alphabetic', 'hanging') The text baseline for the legend labels. \cr
\code{label_text_color} \tab The text color for the legend labels. \cr
\code{label_text_font} \tab The text font for the legend labels. \cr
\code{label_text_font_size} \tab The text font size for the legend labels. \cr
\code{label_text_font_style} \tab ('normal', 'italic', 'bold') The text font style for the legend labels. \cr
\code{label_width} \tab The width (in pixels) of the area that legend labels should occupy. \cr
\code{legend_padding} \tab Amount of padding around the legend. \cr
\code{legend_spacing} \tab Amount of spacing between legend entries. \cr
}
}
\examples{
figure() \%>\%
  ly_points(Sepal.Length, Sepal.Width, color = Species, data = iris) \%>\%
  legend_control(orientation = "top_left", label_width = 2,
    border_line_alpha = 0.5, border_line_cap = "butt")
}

