% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_lines.R
\name{ly_abline}
\alias{ly_abline}
\title{Add an "abline" layer to a Bokeh figure}
\usage{
ly_abline(
  fig,
  a = NULL,
  b = NULL,
  v = NULL,
  h = NULL,
  coef = NULL,
  color = "black",
  alpha = NULL,
  width = 1,
  type = 1,
  legend = NULL,
  lname = NULL,
  lgroup = NULL,
  ...
)
}
\arguments{
\item{fig}{figure to modify}

\item{a, b}{the intercept and slope of the line(s) to draw}

\item{v}{the x value(s) for vertical lines}

\item{h}{the y value(s) for horizontal lines}

\item{coef}{a vector of length two giving the intercept and slope}

\item{color}{color to use to stroke lines with - a hex code (with no alpha) or any of the 147 named CSS colors, e.g 'green', 'indigo'}

\item{alpha}{transparency value for the line between 0 (transparent) and 1 (opaque)}

\item{width}{stroke width in units of pixels}

\item{type}{an integer between 1 and 6 matching the \code{lty} property in \code{\link[graphics]{par}} or an array of integer pixel distances that describe the on-off pattern of dashing to use}

\item{legend}{either a logical specifying not to plot a legend for this layer (FALSE) or a string indicating the name of the legend entry for this layer (note that when mapping plot attributes to variables in \code{data}, a legend is automatically created and does not need to be specified - see "Mapped plot attributes and legends" below)}

\item{lname}{layer name}

\item{lgroup}{layer group}

\item{\ldots}{additional parameters for fine control over line properties (see "Additional parameters" below)}
}
\description{
Draws one or more straight lines.
}
\section{Mapped plot attributes and legends}{
 When specifying an input data frame for a layer through the \code{data} argument, columns of \code{data} can be used to specify various plot attributes such as \code{color}, etc. For example, with \code{ly_points(..., data = iris, color = Species)}, the \code{Species} variable is used to determine how to color the points.  Here, \code{Species} is "mapped" to the \code{color} attribute.  Both continuous and categorical variables can be mapped.  In the case of continuous variables, the range is cut into slices and attributes are applied to each interval.  The mapping from the values of the variable to the actual plot attributes is determined based on the theme.
}

\section{Additional parameters}{

\tabular{ll}{
  \code{line_join} \tab how path segments should be joined together 'miter' 'round' 'bevel' \cr
  \code{line_cap} \tab how path segments should be terminated 'butt' 'round' 'square' \cr
  \code{line_dash} \tab an integer between 1 and 6 matching the \code{lty} property in \code{\link[graphics]{par}} or an array of integer pixel distances that describe the on-off pattern of dashing to use \cr
  \code{line_dash_offset} \tab the distance in pixels into the line_dash that the pattern should start from
 }
}

\examples{
\donttest{
z <- lm(dist ~ speed, data = cars)
p <- figure() \%>\%
  ly_points(cars, hover = cars) \%>\%
  ly_lines(lowess(cars), legend = "lowess") \%>\%
  ly_abline(z, type = 2, legend = "lm", width = 2)
p
}
# abline with mixed axes for h and v
figure() \%>\%
  ly_points(1:26, letters) \%>\%
  ly_abline(h = "j") \%>\%
  ly_abline(v = 10)

# multiple hv lines
figure() \%>\%
  ly_points(1:10) \%>\%
  ly_abline(v = 1:10) \%>\%
  ly_abline(h = 1:10)

# multiple ab lines
figure() \%>\%
  ly_points(0:10) \%>\%
  ly_abline(0, seq(0, 1, by = 0.1))
}
\seealso{
Other layer functions: 
\code{\link{ly_annular_wedge}()},
\code{\link{ly_annulus}()},
\code{\link{ly_arc}()},
\code{\link{ly_bar}()},
\code{\link{ly_bezier}()},
\code{\link{ly_boxplot}()},
\code{\link{ly_contour}()},
\code{\link{ly_crect}()},
\code{\link{ly_curve}()},
\code{\link{ly_density}()},
\code{\link{ly_hist}()},
\code{\link{ly_image_url}()},
\code{\link{ly_image}()},
\code{\link{ly_lines}()},
\code{\link{ly_map}()},
\code{\link{ly_multi_line}()},
\code{\link{ly_oval}()},
\code{\link{ly_patch}()},
\code{\link{ly_points}()},
\code{\link{ly_polygons}()},
\code{\link{ly_quadratic}()},
\code{\link{ly_quantile}()},
\code{\link{ly_ray}()},
\code{\link{ly_rect}()},
\code{\link{ly_segments}()},
\code{\link{ly_text}()},
\code{\link{ly_wedge}()}
}
\concept{layer functions}
