% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_base.R
\name{ly_annular_wedge}
\alias{ly_annular_wedge}
\title{Add an "annular_wedge" layer to a Bokeh figure}
\usage{
ly_annular_wedge(fig, x, y = NULL, data = figure_data(fig),
  inner_radius = 0.1, outer_radius = 0.3, start_angle = 0, end_angle = 2
  * pi, direction = "anticlock", color = NULL, alpha = NULL,
  hov_color = NULL, hov_alpha = NULL, ns_color = NULL, ns_alpha = NULL,
  sel_color = NULL, sel_alpha = NULL, hover = NULL, url = NULL,
  legend = TRUE, lname = NULL, lgroup = NULL, ...)
}
\arguments{
\item{fig}{figure to modify}

\item{x}{values or field name of center x coordinates}

\item{y}{values or field name of center y coordinates}

\item{data}{an optional data frame, providing the source for inputs x, y, and other glyph properties}

\item{inner_radius}{values or field name of inner radii}

\item{outer_radius}{values or field name of outer radii}

\item{start_angle}{the angles to start the annular wedges, in radians, as measured from the horizontal}

\item{end_angle}{the angles to end the annular wedges, in radians, as measured from the horizontal}

\item{direction}{direction to turn between starting and ending angles ("anticlock", "clock")}
}
\description{
Add an "annular_wedge" layer to a Bokeh figure
}
\examples{
rescale <- function(x)
  (x - min(x)) / diff(range(x))
figure() \%>\%
  ly_annular_wedge(Sepal.Length, Sepal.Width, data = iris,
    end_angle = rescale(Petal.Length) * 2 * pi, color = Species,
    inner_radius = 0.1, outer_radius = 0.15, alpha = 0.5,
    hover = Species)
}
\seealso{
Other layer functions: \code{\link{ly_abline}},
  \code{\link{ly_annulus}}, \code{\link{ly_arc}},
  \code{\link{ly_bar}}, \code{\link{ly_bezier}},
  \code{\link{ly_boxplot}}, \code{\link{ly_contour}},
  \code{\link{ly_crect}}, \code{\link{ly_curve}},
  \code{\link{ly_density}}, \code{\link{ly_hist}},
  \code{\link{ly_image_url}}, \code{\link{ly_image}},
  \code{\link{ly_lines}}, \code{\link{ly_multi_line}},
  \code{\link{ly_oval}}, \code{\link{ly_patch}},
  \code{\link{ly_points}}, \code{\link{ly_polygons}},
  \code{\link{ly_quadratic}}, \code{\link{ly_quantile}},
  \code{\link{ly_ray}}, \code{\link{ly_rect}},
  \code{\link{ly_segments}}, \code{\link{ly_text}},
  \code{\link{ly_wedge}}
}
