% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_base_plus.R
\name{ly_contour}
\alias{ly_contour}
\title{Add a "contour" layer to a Bokeh figure}
\usage{
ly_contour(fig, z, x = seq(0, 1, length.out = nrow(z)), y = seq(0, 1,
  length.out = ncol(z)), nlevels = 10, levels = pretty(range(z, na.rm =
  TRUE), nlevels), color = "black", alpha = 1, width = 1, type = 1,
  hov_color = NULL, hov_alpha = NULL, ns_color = NULL, ns_alpha = NULL,
  lname = NULL, lgroup = NULL, ...)
}
\arguments{
\item{fig}{Figure to modify.}

\item{z}{a matrix containing the values to compute contour lines for}

\item{x, y}{locations of grid lines at which the values in \code{image} are measured (see \code{\link[grDevices]{contourLines}})}

\item{nlevels, levels}{parameters sent to \code{\link[grDevices]{contourLines}})}
}
\description{
Computes and draws contour lines.
}
\seealso{
Other layer functions: \code{\link{ly_annular_wedge}},
  \code{\link{ly_annulus}}, \code{\link{ly_arc}},
  \code{\link{ly_bar}}, \code{\link{ly_bezier}},
  \code{\link{ly_boxplot}}, \code{\link{ly_crect}},
  \code{\link{ly_curve}}, \code{\link{ly_density}},
  \code{\link{ly_hist}}, \code{\link{ly_image_url}},
  \code{\link{ly_image}}, \code{\link{ly_lines}},
  \code{\link{ly_multi_line}}, \code{\link{ly_oval}},
  \code{\link{ly_patch}}, \code{\link{ly_points}},
  \code{\link{ly_polygons}}, \code{\link{ly_quadratic}},
  \code{\link{ly_quantile}}, \code{\link{ly_ray}},
  \code{\link{ly_rect}}, \code{\link{ly_segments}},
  \code{\link{ly_text}}, \code{\link{ly_wedge}}
}
