% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_stat.R
\name{ly_density}
\alias{ly_density}
\title{Add a "density" layer to a Bokeh figure}
\usage{
ly_density(fig, x, data = figure_data(fig), bw = "nrd0", adjust = 1,
  kernel = c("gaussian", "epanechnikov", "rectangular", "triangular",
  "biweight", "cosine", "optcosine"), weights = NULL, window = kernel,
  n = 512, cut = 3, na.rm = FALSE, color = "black", alpha = 1,
  width = 1, type = 1, hov_color = NULL, hov_alpha = NULL,
  ns_color = NULL, ns_alpha = NULL, sel_color = NULL, sel_alpha = NULL,
  legend = NULL, lname = NULL, lgroup = NULL, ...)
}
\arguments{
\item{fig}{Figure to modify.}

\item{x, bw, adjust, kernel, weights, window, n, cut, na.rm}{parameters passed to \code{\link[stats]{density}}}

\item{data}{an optional data frame, providing the source for x}

\item{legend}{text to display in the legend entry for the density line}
}
\description{
Draws a kernel density estimate
}
\examples{
h <- figure(width = 600, height = 400) \%>\%
  ly_hist(eruptions, data = faithful, breaks = 40, freq = FALSE) \%>\%
  ly_density(eruptions, data = faithful)
h
}
\seealso{
Other layer functions: \code{\link{ly_annular_wedge}},
  \code{\link{ly_annulus}}, \code{\link{ly_arc}},
  \code{\link{ly_bar}}, \code{\link{ly_bezier}},
  \code{\link{ly_boxplot}}, \code{\link{ly_contour}},
  \code{\link{ly_crect}}, \code{\link{ly_curve}},
  \code{\link{ly_hist}}, \code{\link{ly_image_url}},
  \code{\link{ly_image}}, \code{\link{ly_lines}},
  \code{\link{ly_multi_line}}, \code{\link{ly_oval}},
  \code{\link{ly_patch}}, \code{\link{ly_points}},
  \code{\link{ly_polygons}}, \code{\link{ly_quadratic}},
  \code{\link{ly_quantile}}, \code{\link{ly_ray}},
  \code{\link{ly_rect}}, \code{\link{ly_segments}},
  \code{\link{ly_text}}, \code{\link{ly_wedge}}
}
