% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/layer_hexbin.R
\name{ly_hexbin}
\alias{ly_hexbin}
\title{Add a "hexbin" layer to a Bokeh figure}
\usage{
ly_hexbin(fig, x, y, data = NULL, xbins = 30, shape = 1,
  style = "colorscale", trans = NULL, inv = NULL, palette = "RdYlGn11",
  fill = TRUE, line = FALSE, hover = TRUE)
}
\arguments{
\item{fig}{figure to modify}

\item{x}{values or field name of center x coordinates to be binned}

\item{y}{values or field name of center y coordinates to be binned}

\item{data}{an optional data frame, providing the source for x and y}

\item{xbins,shape}{parameters passed to \code{\link[hexbin]{hexbin}}}

\item{style}{type of plotting for hexbins (see \code{\link[hexbin]{grid.hexagons}}) - "colorramp" and "lattice" are currently supported}

\item{trans,inv}{transformation and inverse transformation function for the bin counts}

\item{palette}{name of color palette to use for color ramp (see \href{http://bokeh.pydata.org/en/latest/docs/reference/palettes.html}{here} for acceptable values)}

\item{fill}{logical - should hexagons be filled?}

\item{line}{logical - should hexagons have an outline?}

\item{hover}{logical - should a hover tool be added to show the count in each hexagon?}
}
\description{
Add a "hexbin" layer to a Bokeh figure
}

