% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_base.R
\name{ly_image_url}
\alias{ly_image_url}
\title{Add an "image_url" layer to a Bokeh figure}
\usage{
ly_image_url(fig, x = NULL, y = NULL, url, w = 10, h = 10,
  data = figure_data(fig), dilate = TRUE, anchor = "top_left",
  angle = 0, w_units = "screen", h_units = "screen", global_alpha = 1,
  retry_attempts = 0, retry_timeout = 0, lgroup = NULL, lname = NULL,
  ...)
}
\arguments{
\item{fig}{figure to modify}

\item{x}{x coordinates}

\item{y}{y coordinates}

\item{w, h}{values or field names of width and height of image}

\item{data}{an optional data frame, providing the source for inputs x, y, and other properties}

\item{dilate}{logical - Whether to always round fractional pixel locations in such a way as to make the images bigger. This setting may be useful if pixel rounding errors are causing images to have a gap between them, when they should appear flush.}

\item{anchor}{Where the image is anchored to with respect to \code{x} and \code{y}. One of 'top_left', 'top_center', 'top_right', 'center_left', 'center', 'center_right', 'bottom_left', 'bottom_center', 'bottom_right'.}

\item{angle}{values or field name of the angle to rotate the image, in radians}

\item{global_alpha}{An overall opacity that each image is rendered with (in addition to any inherent alpha values in the image itself).}

\item{retry_attempts}{Number of attempts to retry loading the images from the specified URL.}

\item{retry_timeout}{Timeout (in ms) between retry attempts to load the image from the specified URL.}

\item{image_url}{values or field name of image URLs}
}
\description{
Renders raster images from URLs at provided coordinates
}
\seealso{
Other layer functions: \code{\link{ly_abline}},
  \code{\link{ly_annular_wedge}}, \code{\link{ly_annulus}},
  \code{\link{ly_arc}}, \code{\link{ly_bar}},
  \code{\link{ly_bezier}}, \code{\link{ly_boxplot}},
  \code{\link{ly_contour}}, \code{\link{ly_crect}},
  \code{\link{ly_curve}}, \code{\link{ly_density}},
  \code{\link{ly_hist}}, \code{\link{ly_image}},
  \code{\link{ly_lines}}, \code{\link{ly_multi_line}},
  \code{\link{ly_oval}}, \code{\link{ly_patch}},
  \code{\link{ly_points}}, \code{\link{ly_polygons}},
  \code{\link{ly_quadratic}}, \code{\link{ly_quantile}},
  \code{\link{ly_ray}}, \code{\link{ly_rect}},
  \code{\link{ly_segments}}, \code{\link{ly_text}},
  \code{\link{ly_wedge}}
}
