% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_raw.R
\name{ly_lines}
\alias{ly_lines}
\title{Add a "lines" layer to a Bokeh figure
Draws lines with the given coordinates.}
\usage{
ly_lines(fig, x = NULL, y = NULL, data = figure_data(fig), type = 1,
  width = 1, color = NULL, alpha = NULL, hover = NULL, legend = TRUE,
  hov_color = NULL, hov_alpha = NULL, ns_color = NULL, ns_alpha = NULL,
  lgroup = NULL, lname = NULL, ...)
}
\arguments{
\item{fig}{figure to modify}

\item{x}{values or field name of line x coordinates}

\item{y}{values or field name of line y coordinates}

\item{data}{an optional data frame, providing the source for inputs x, y, and other glyph properties}

\item{group}{values or field name of a grouping variable to break lines up by}
}
\description{
Add a "lines" layer to a Bokeh figure
Draws lines with the given coordinates.
}
