% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_base.R
\name{ly_oval}
\alias{ly_oval}
\title{Add an "oval" layer to a Bokeh figure}
\usage{
ly_oval(fig, x, y = NULL, width = 1, height = 1, angle = 0,
  data = figure_data(fig), color = NULL, alpha = NULL, hov_color = NULL,
  hov_alpha = NULL, ns_color = NULL, ns_alpha = NULL, sel_color = NULL,
  sel_alpha = NULL, hover = NULL, url = NULL, lname = NULL,
  lgroup = NULL, ...)
}
\arguments{
\item{fig}{Figure to modify.}

\item{x}{values or field name of center x coordinates}

\item{y}{values or field name of center y coordinates}

\item{width}{values or field name of widths}

\item{height}{values or field name of heights}

\item{angle}{values or field name of rotation angles}

\item{data}{an optional data frame, providing the source for inputs x, y, and other glyph properties}
}
\description{
Add an "oval" layer to a Bokeh figure
}
\seealso{
Other layer functions: \code{\link{ly_annular_wedge}},
  \code{\link{ly_annulus}}, \code{\link{ly_arc}},
  \code{\link{ly_bar}}, \code{\link{ly_bezier}},
  \code{\link{ly_boxplot}}, \code{\link{ly_contour}},
  \code{\link{ly_crect}}, \code{\link{ly_curve}},
  \code{\link{ly_density}}, \code{\link{ly_hist}},
  \code{\link{ly_image_url}}, \code{\link{ly_image}},
  \code{\link{ly_lines}}, \code{\link{ly_multi_line}},
  \code{\link{ly_patch}}, \code{\link{ly_points}},
  \code{\link{ly_polygons}}, \code{\link{ly_quadratic}},
  \code{\link{ly_quantile}}, \code{\link{ly_ray}},
  \code{\link{ly_rect}}, \code{\link{ly_segments}},
  \code{\link{ly_text}}, \code{\link{ly_wedge}}
}
