% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_base.R
\name{ly_patch}
\alias{ly_patch}
\title{Add a "patch" layer to a Bokeh figure}
\usage{
ly_patch(fig, x, y = NULL, data = figure_data(fig), color = NULL,
  alpha = NULL, hov_color = NULL, hov_alpha = NULL, ns_color = NULL,
  ns_alpha = NULL, sel_color = NULL, sel_alpha = NULL, hover = NULL,
  url = NULL, legend = TRUE, lname = NULL, lgroup = NULL, ...)
}
\arguments{
\item{fig}{figure to modify}

\item{x}{values or field name of patch x coordinates}

\item{y}{values or field name of patch y coordinates}

\item{data}{an optional data frame, providing the source for inputs x, y, and other glyph properties}
}
\description{
Add a "patch" layer to a Bokeh figure
}
\note{
This function is included for completeness as it maps to Bokeh's \code{patch} glyph, but the same and more functionality can be obtained with \code{\link{ly_polygons}}.
}
\seealso{
Other layer functions: \code{\link{ly_crect}},
  \code{\link{ly_lines}}, \code{\link{ly_oval}},
  \code{\link{ly_points}}, \code{\link{ly_rect}},
  \code{\link{ly_text}}
}
