% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_raw.R
\name{ly_points}
\alias{ly_points}
\title{Add a "points" layer to a Bokeh figure
Draws points with the given coordinates.}
\usage{
ly_points(fig, x = NULL, y = NULL, data = figure_data(fig), glyph = 21,
  color = NULL, alpha = NULL, size = NULL, hover = NULL,
  legend = TRUE, hov_color = NULL, hov_alpha = NULL, ns_color = NULL,
  ns_alpha = NULL, lgroup = NULL, lname = NULL, ...)
}
\arguments{
\item{fig}{figure to modify}

\item{x}{values or field name of line x coordinates}

\item{y}{values or field name of line y coordinates}

\item{data}{an optional data frame, providing the source for inputs x, y, and other glyph properties}
}
\description{
Add a "points" layer to a Bokeh figure
Draws points with the given coordinates.
}
\seealso{
Other layer functions: \code{\link{ly_lines}}
}
