% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_base.R
\name{ly_polygons}
\alias{ly_polygons}
\title{Add a "polygons" layer to a Bokeh figure}
\usage{
ly_polygons(fig, xs = NULL, ys = NULL, data = figure_data(fig),
  type = 1, width = 1, color = NULL, alpha = NULL, hover = NULL,
  legend = TRUE, hov_color = NULL, hov_alpha = NULL, ns_color = NULL,
  ns_alpha = NULL, lgroup = NULL, lname = NULL, ...)
}
\arguments{
\item{fig}{Figure to modify.}

\item{xs}{vector or list of values or field name of polygon x coordinates - see details}

\item{ys}{vector or list of values or field name of polygon y coordinates - see details}

\item{data}{an optional data frame, providing the source for inputs xs, ys, group, and other glyph properties}

\item{group}{vector or field name of grouping variable - see details}
}
\description{
Add a "polygons" layer to a Bokeh figure
}
\details{
\code{xs} and \code{ys} can be a list of vectors, each element for one polygon to be drawn, or can be vectors with the \code{group} argument specifying how to break them up into individual polygons.
}
\seealso{
Other layer functions: \code{\link{ly_abline}},
  \code{\link{ly_annular_wedge}}, \code{\link{ly_annulus}},
  \code{\link{ly_arc}}, \code{\link{ly_bar}},
  \code{\link{ly_bezier}}, \code{\link{ly_boxplot}},
  \code{\link{ly_contour}}, \code{\link{ly_crect}},
  \code{\link{ly_curve}}, \code{\link{ly_density}},
  \code{\link{ly_hist}}, \code{\link{ly_image_url}},
  \code{\link{ly_image}}, \code{\link{ly_lines}},
  \code{\link{ly_multi_line}}, \code{\link{ly_oval}},
  \code{\link{ly_patch}}, \code{\link{ly_points}},
  \code{\link{ly_quadratic}}, \code{\link{ly_quantile}},
  \code{\link{ly_ray}}, \code{\link{ly_rect}},
  \code{\link{ly_segments}}, \code{\link{ly_text}},
  \code{\link{ly_wedge}}
}
