% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_base.R
\name{ly_rect}
\alias{ly_rect}
\title{Add a "rect" layer to a Bokeh figure}
\usage{
ly_rect(fig, xleft, ybottom, xright, ytop, data = figure_data(fig),
  color = NULL, alpha = NULL, hov_color = NULL, hov_alpha = NULL,
  ns_color = NULL, ns_alpha = NULL, sel_color = NULL, sel_alpha = NULL,
  hover = NULL, url = NULL, legend = TRUE, lname = NULL,
  lgroup = NULL, ...)
}
\arguments{
\item{fig}{figure to modify}

\item{xleft}{values or field name of left edges}

\item{ybottom}{values or field name of bottom edges}

\item{xright}{values or field name of right edges}

\item{ytop}{values or field name of top edges}

\item{data}{an optional data frame, providing the source for inputs xleft, ybottom, xright, ytop, and other glyph properties}
}
\description{
Add a "rect" layer to a Bokeh figure
}
\seealso{
Other layer functions: \code{\link{ly_crect}},
  \code{\link{ly_lines}}, \code{\link{ly_oval}},
  \code{\link{ly_patch}}, \code{\link{ly_points}},
  \code{\link{ly_text}}
}
