% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_base.R
\name{ly_text}
\alias{ly_text}
\title{Add a "text" layer to a Bokeh figure}
\usage{
ly_text(fig, x, y = NULL, text = NULL, color = NULL, alpha = NULL,
  angle = NULL, data = figure_data(fig), x_offset = NULL,
  y_offset = NULL, font = NULL, font_size = NULL, font_style = NULL,
  align = NULL, baseline = NULL, hov_color = NULL, hov_alpha = NULL,
  ns_color = NULL, ns_alpha = NULL, lgroup = NULL, lname = NULL, ...)
}
\arguments{
\item{fig}{figure to modify}

\item{x}{x coordinates of text anchors}

\item{y}{y coordinates of text anchors}

\item{text}{text values to render}

\item{color}{text color values for the text}

\item{alpha}{text alpha values for the text}

\item{angle}{angle to rotate the text in radians}

\item{data}{an optional data frame, providing the source for inputs x, y, text, and other glyph properties}

\item{x_offset}{offset values to apply to the x-coordinates}

\item{y_offset}{offset values to apply to the y-coordinates}

\item{font}{text font values for the text}

\item{font_size}{text font size values for the text}

\item{font_style}{text font style values for the text ("normal", "italic", "bold")}

\item{align}{text align values for the text ("left", "right", "center")}

\item{baseline}{text baseline values for the text ("top", "middle", "bottom", "alphabetic", "hanging")}
}
\description{
Add a "text" layer to a Bokeh figure
}
\seealso{
Other layer functions: \code{\link{ly_crect}},
  \code{\link{ly_lines}}, \code{\link{ly_oval}},
  \code{\link{ly_patch}}, \code{\link{ly_points}},
  \code{\link{ly_rect}}
}
