% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_palettes.R
\name{pal_color}
\alias{pal_color}
\alias{pal_tableau}
\alias{pal_bk_glyph}
\alias{pal_gradient}
\alias{pal_bk_line_dash}
\title{Palettes for themes.}
\usage{
pal_color(colors)

pal_tableau(pal = "Tableau10")

pal_bk_glyph()

pal_gradient(cols = c("#66C2A4", "#41AE76", "#238B45", "#006D2C", "#00441B"),
  space = "rgb")

pal_bk_line_dash()
}
\arguments{
\item{colors}{A vector of colors to be used in the color palette.}

\item{pal}{Palette name.}

\item{cols}{A vector of colors to ramp across for a continuous palette.}

\item{space}{Passed on to \code{\link{colorRampPalette}[grDevices]}.}

\item{min}{Minimum value.}

\item{max}{Maximum value.}
}
\description{
Palettes for themes.

Palettes for themes.
}
