% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_palettes.R
\name{pal_tableau}
\alias{pal_bk_glyph}
\alias{pal_bk_line_dash}
\alias{pal_bk_line_width}
\alias{pal_gradient}
\alias{pal_size}
\alias{pal_tableau}
\title{Palettes for themes}
\usage{
pal_tableau(pal = "Tableau10")

pal_bk_glyph()

pal_gradient()

pal_size(min = 2, max = 20)

pal_bk_line_dash()

pal_bk_line_width()
}
\arguments{
\item{pal}{palette name}

\item{min}{minimum value}

\item{max}{maximum value}
}
\description{
Palettes for themes
}

