% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_palettes.R
\name{pal_color}
\alias{pal_bk_glyph}
\alias{pal_bk_line_dash}
\alias{pal_bk_line_width}
\alias{pal_color}
\alias{pal_gradient}
\alias{pal_size}
\alias{pal_tableau}
\title{Palettes for themes}
\usage{
pal_color(colors)

pal_tableau(pal = "Tableau10")

pal_bk_glyph()

pal_gradient(cols = c("#66C2A4", "#41AE76", "#238B45", "#006D2C", "#00441B"),
  space = "rgb")

pal_size(min = 2, max = 20)

pal_bk_line_dash()

pal_bk_line_width()
}
\arguments{
\item{colors}{a vector of colors to be used in the color palette}

\item{pal}{palette name}

\item{cols}{a vector of colors to ramp across for a continuous palette}

\item{space}{passed on to \code{\link{colorRampPalette}[grDevices]}}

\item{min}{minimum value}

\item{max}{maximum value}
}
\description{
Palettes for themes

Palettes for themes
}

