% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_gist.R
\name{rbokeh2gist}
\alias{rbokeh2gist}
\title{Export rbokeh plot to a gist}
\usage{
rbokeh2gist(fig, name, created = NULL, description = "")
}
\arguments{
\item{fig}{a string containing R code to create an rbokeh figure}

\item{name}{name of the gist}

\item{created}{optional string for a "Created by" to preceed the README}

\item{description}{text to go in README.md}
}
\description{
Export rbokeh plot to a gist
}
\note{
This requires that you have a github personal access token stored as an environment variable \code{GITHUB_PAT}.  See \code{\link{gist_create}} for more information.

Also note that this currently can't handle thumbnails but will find a way to do that soon.
}

