% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer__utils.R
\name{sub_names}
\alias{sub_names}
\title{Retrieve and properly parse all data}
\usage{
sub_names(fig, data, arg_obj, process_data_and_names = TRUE)
}
\arguments{
\item{fig}{figure to be used}

\item{data}{data to be used}

\item{arg_obj}{args object supplied by \code{grab}}

\item{process_data_and_names}{boolean to determine if the data and x_name and y_name should be post processed}
}
\value{
list of three groups: data, info, and params
}
\description{
Retrieve and properly parse all data
}

