% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_tools.R
\name{tool_crosshair}
\alias{tool_crosshair}
\title{Add "crosshair" tool to a Bokeh figure}
\usage{
tool_crosshair(fig, dimensions = NULL, line_width = NULL,
  line_alpha = NULL, line_color = NULL, remove = FALSE)
}
\arguments{
\item{dimensions}{Which dimensions the crosshair tool is to track. By default, both a vertical and horizontal line will be drawn. If only "width" is supplied, only a horizontal line will be drawn. If only "height" is supplied, only a vertical line will be drawn. One of 'width', 'height', or 'both'.}

\item{line_width}{Stroke width in units of pixels.}

\item{line_alpha}{An alpha value to use to stroke paths with.}

\item{line_color}{A color to use to stroke the crosshair paths with.}
}
\description{
Add "crosshair" tool to a Bokeh figure
}
