% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_tools.R
\name{tool_help}
\alias{tool_help}
\title{Add "help" tool to a Bokeh figure}
\usage{
tool_help(fig, redirect = NULL, help_tooltip = NULL, remove = FALSE)
}
\arguments{
\item{redirect}{Site to be redirected through upon click.}

\item{help_tooltip}{Tooltip displayed when hovering over the help icon.}
}
\description{
Add "help" tool to a Bokeh figure
}
