% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_tools.R
\name{tool_lasso_select}
\alias{tool_lasso_select}
\title{Add "lasso_select" tool to a Bokeh figure}
\usage{
tool_lasso_select(fig, select_every_mousemove = TRUE, line_color = NULL,
  line_alpha = NULL, fill_color = NULL, fill_alpha = NULL,
  line_width = NULL, line_dash = NULL, line_dash_offset = NULL,
  line_join = NULL, line_cap = NULL, names = NULL, callback = NULL,
  remove = FALSE)
}
\arguments{
\item{fig}{Figure to modify.}

\item{select_every_mousemove}{Whether a selection computation should happen on every mouse event, or only once, when the selection region is completed. Default: True}

\item{line_color}{The line color values for the polygon.}

\item{line_alpha}{The line alpha values for the polygon.}

\item{fill_color}{The fill color values for the polygon.}

\item{fill_alpha}{The fill alpha values for the polygon.}

\item{line_width}{The line width values for the polygon.}

\item{line_dash}{The line dash values for the polygon.}

\item{line_dash_offset}{The line dash offset values for the polygon.}

\item{line_join}{The line join values for the polygon.}

\item{line_cap}{The line cap values for the polygon.}

\item{names}{A list of names to query for. If set, only renderers that have a matching value for their "name" attribute will be used.}

\item{callback}{A callback to run in the browser on every selection of a lasso area. The cb_data parameter that is available to the Callback code will contain one LassoSelectTool-specific field: :geometry: object containing the coordinates of the lasso area.}

\item{remove}{Logical indicating whether the tool should be removed.}
}
\description{
Add "lasso_select" tool to a Bokeh figure
}
\examples{
figure(tools = NULL) \%>\%
  ly_points(1:10) \%>\%
  tool_lasso_select(line_color = "red", callback = "console.log('calling back...')")
}
\seealso{
Other tools: \code{\link{tool_box_select}},
  \code{\link{tool_box_zoom}},
  \code{\link{tool_crosshair}}, \code{\link{tool_help}},
  \code{\link{tool_pan}}, \code{\link{tool_poly_select}},
  \code{\link{tool_redo}}, \code{\link{tool_reset}},
  \code{\link{tool_save}}, \code{\link{tool_undo}},
  \code{\link{tool_wheel_pan}},
  \code{\link{tool_wheel_zoom}},
  \code{\link{tool_zoom_in}}, \code{\link{tool_zoom_out}}
}
