% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_tools.R
\name{tool_pan}
\alias{tool_pan}
\title{Add "pan" tool to a Bokeh figure}
\usage{
tool_pan(fig, dimensions = NULL, remove = FALSE)
}
\arguments{
\item{dimensions}{a vector specifying whether the pan tool should pan with respect to the x axis ("width") and the y axis ("height") or both (c("width", "height"))}
}
\description{
Add "pan" tool to a Bokeh figure
}
\examples{
\donttest{
# only pan on x axis
figure() \%>\% ly_points(1:10) \%>\%
 tool_pan(dimensions = "height")
}
}
