% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_tools.R
\name{tool_poly_select}
\alias{tool_poly_select}
\title{Add "poly_select" tool to a Bokeh figure}
\usage{
tool_poly_select(fig, line_color = NULL, line_alpha = NULL,
  fill_color = NULL, fill_alpha = NULL, line_width = NULL,
  line_dash = NULL, line_dash_offset = NULL, line_join = NULL,
  line_cap = NULL, names = NULL, remove = FALSE)
}
\arguments{
\item{names}{A list of names to query for. If set, only renderers that have a matching value for their ``name`` attribute will be used.}
}
\description{
Add "poly_select" tool to a Bokeh figure
}
