% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_tools.R
\name{tool_redo}
\alias{tool_redo}
\title{Add "redo" tool to a Bokeh figure}
\usage{
tool_redo(fig, remove = FALSE)
}
\arguments{
\item{fig}{Figure to modify.}

\item{remove}{Logical indicating whether the tool should be removed.}
}
\description{
Add "redo" tool to a Bokeh figure
}
\examples{
figure() \%>\%
  ly_points(1:10) \%>\%
  tool_undo() \%>\%
  tool_redo()
}
