% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_tools.R
\name{tool_reset}
\alias{tool_reset}
\title{Add "reset" tool to a Bokeh figure}
\usage{
tool_reset(fig, reset_size = TRUE, remove = FALSE)
}
\arguments{
\item{fig}{Figure to modify.}

\item{reset_size}{Whether activating the Reset tool should also reset the plot's canvas dimensions to their original size.}

\item{remove}{Logical indicating whether the tool should be removed.}
}
\description{
Add "reset" tool to a Bokeh figure
}
\examples{
figure() \%>\%
  ly_points(1:10) \%>\%
  tool_reset()
}
\seealso{
Other tools: \code{\link{tool_box_select}},
  \code{\link{tool_box_zoom}},
  \code{\link{tool_crosshair}}, \code{\link{tool_help}},
  \code{\link{tool_lasso_select}}, \code{\link{tool_pan}},
  \code{\link{tool_poly_select}}, \code{\link{tool_redo}},
  \code{\link{tool_save}}, \code{\link{tool_undo}},
  \code{\link{tool_wheel_pan}},
  \code{\link{tool_wheel_zoom}},
  \code{\link{tool_zoom_in}}, \code{\link{tool_zoom_out}}
}
