% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_tools.R
\name{tool_wheel_zoom}
\alias{tool_wheel_zoom}
\title{Add "wheel_zoom" tool to a Bokeh figure}
\usage{
tool_wheel_zoom(fig, dimensions = NULL, remove = FALSE)
}
\arguments{
\item{fig}{Figure to modify.}

\item{dimensions}{Which dimensions the wheel zoom tool is constrained to act in. By default the wheel zoom tool will zoom in any dimension, but can be configured to only zoom horizontally across the width of the plot, or vertically across the height of the plot. One of 'width', 'height', or 'both'.}

\item{remove}{Logical indicating whether the tool should be removed.}
}
\description{
Add "wheel_zoom" tool to a Bokeh figure
}
\examples{
figure(tools = NULL) \%>\%
  ly_points(1:10) \%>\%
  tool_wheel_zoom()
}
\seealso{
Other tools: \code{\link{tool_box_select}},
  \code{\link{tool_box_zoom}},
  \code{\link{tool_crosshair}}, \code{\link{tool_help}},
  \code{\link{tool_lasso_select}}, \code{\link{tool_pan}},
  \code{\link{tool_poly_select}}, \code{\link{tool_redo}},
  \code{\link{tool_reset}}, \code{\link{tool_save}},
  \code{\link{tool_undo}}, \code{\link{tool_wheel_pan}},
  \code{\link{tool_zoom_in}}, \code{\link{tool_zoom_out}}
}
