% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fig_axis.R
\name{x_axis}
\alias{x_axis}
\title{Add x axis to a Bokeh figure}
\usage{
x_axis(fig, label, position = "below", log = FALSE, grid = TRUE,
  num_minor_ticks = 5)
}
\arguments{
\item{fig}{figure to modify}

\item{label}{axis label}

\item{position}{where to place the axis (either "above" or "below")}

\item{log}{logical - should a log axis be used?}

\item{grid}{logical - should a reference grid be shown for this axis?}

\item{num_minor_ticks}{number of minor ticks}
}
\description{
Add x axis to a Bokeh figure
}
\examples{
\donttest{
figure() \%>\%
  ly_points(rexp(1000), rexp(1000)) \%>\%
  x_axis(label = "x", log = TRUE) \%>\%
  y_axis(label = "y", log = TRUE)
}
}
\seealso{
Other axes: \code{\link{y_axis}}
}

