% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fig_axis.R
\name{y_axis}
\alias{y_axis}
\title{Add y axis to a Bokeh figure}
\usage{
y_axis(fig, label, position = "left", log = FALSE, grid = TRUE,
  num_minor_ticks = 5, visible = TRUE, ...)
}
\arguments{
\item{fig}{figure to modify}

\item{label}{axis label}

\item{position}{where to place the axis (either "left" or "right")}

\item{log}{logical - should a log axis be used?}

\item{grid}{logical - should a reference grid be shown for this axis?}

\item{num_minor_ticks}{number of minor ticks}

\item{visible}{should axis be shown?}

\item{\ldots}{additional parameters for fine control over axes (see "Additional parameters" below)}
}
\description{
Add y axis to a Bokeh figure
}
\section{Additional parameters}{

\tabular{ll}{
\code{axis_label_standoff} \tab (integer) The distance in pixels that the axis labels should be offset from the tick labels. \cr
\code{axis_label_text_align} \tab ('left', 'right', 'center') The text align of the axis label. \cr
\code{axis_label_text_alpha} \tab (numeric) The text alpha of the axis label. \cr
\code{axis_label_text_baseline} \tab ('top', 'middle', 'bottom', 'alphabetic', 'hanging') The text baseline of the axis label. \cr
\code{axis_label_text_color} \tab (color) The text color of the axis label. \cr
\code{axis_label_text_font} \tab (string) The text font of the axis label. \cr
\code{axis_label_text_font_size} \tab (string - e.g. '12pt') The text font size of the axis label. \cr
\code{axis_label_text_font_style} \tab ('normal', 'italic', 'bold') The text font style of the axis label. \cr
\code{axis_line_alpha} \tab (numeric) The line alpha of the axis line. \cr
\code{axis_line_cap} \tab ('butt', 'round', 'square') The line cap of the axis line. \cr
\code{axis_line_color} \tab (color) The line color of the axis line. \cr
\code{axis_line_dash} \tab The line dash of the axis line. \cr
\code{axis_line_dash_offset} \tab (integer) The line dash offset of the axis line. \cr
\code{axis_line_join} \tab ('miter', 'round', 'bevel') The line join of the axis line. \cr
\code{axis_line_width} \tab (integer) The line width of the axis line. \cr
\code{major_label_orientation} \tab ('horizontal', 'vertical', or angle in degrees) What direction the major label text should be oriented. If a number is supplied, the angle of the text is measured from horizontal. \cr
\code{major_label_standoff} \tab (integer) The distance in pixels that the major tick labels should be offset from the associated ticks. \cr
\code{major_label_text_align} \tab ('left', 'right', 'center') The text align of the major tick labels. \cr
\code{major_label_text_alpha} \tab (numeric) The text alpha of the major tick labels. \cr
\code{major_label_text_baseline} \tab ('top', 'middle', 'bottom', 'alphabetic', 'hanging') The text baseline of the major tick labels. \cr
\code{major_label_text_color} \tab (color) The text color of the major tick labels. \cr
\code{major_label_text_font} \tab (string - 'Helvetica') The text font of the major tick labels. \cr
\code{major_label_text_font_size} \tab (string - e.g. '12pt') The text font size of the major tick labels. \cr
\code{major_label_text_font_style} \tab ('normal', 'italic', 'bold') The text font style of the major tick labels. \cr
\code{major_tick_in} \tab (integer) The distance in pixels that major ticks should extend into the main plot area. \cr
\code{major_tick_line_alpha} \tab (numeric) The line alpha of the major ticks. \cr
\code{major_tick_line_cap} \tab ('butt', 'round', 'square') The line cap of the major ticks. \cr
\code{major_tick_line_color} \tab (color) The line color of the major ticks. \cr
\code{major_tick_line_dash} \tab The line dash of the major ticks. \cr
\code{major_tick_line_dash_offset} \tab (integer) The line dash offset of the major ticks. \cr
\code{major_tick_line_join} \tab ('miter', 'round', 'bevel') The line join of the major ticks. \cr
\code{major_tick_line_width} \tab (integer) The line width of the major ticks. \cr
\code{major_tick_out} \tab (integer) The distance in pixels that major ticks should extend out of the main plot area. \cr
\code{minor_tick_in} \tab (integer) The distance in pixels that minor ticks should extend into the main plot area. \cr
\code{minor_tick_line_alpha} \tab (numeric) The line alpha of the minor ticks. \cr
\code{minor_tick_line_cap} \tab ('butt', 'round', 'square') The line cap of the minor ticks. \cr
\code{minor_tick_line_color} \tab (color) The line color of the minor ticks. \cr
\code{minor_tick_line_dash} \tab The line dash of the minor ticks. \cr
\code{minor_tick_line_dash_offset} \tab (integer) The line dash offset of the minor ticks. \cr
\code{minor_tick_line_join} \tab ('miter', 'round', 'bevel') The line join of the minor ticks. \cr
\code{minor_tick_line_width} \tab (integer) The line width of the minor ticks. \cr
\code{minor_tick_out} \tab (integer) The distance in pixels that major ticks should extend out of the main plot area. \cr
 }
}
\examples{
\donttest{
figure() \%>\%
  ly_points(rexp(1000), rexp(1000)) \%>\%
  x_axis(label = "x", log = TRUE) \%>\%
  y_axis(label = "y", log = TRUE)
}
}
\seealso{
Other axes: \code{\link{x_axis}}
}

