% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fig_axis.R
\name{y_axis}
\alias{y_axis}
\title{Add y axis to a Bokeh figure}
\usage{
y_axis(fig, label, position = "left", log = FALSE, grid = TRUE,
  num_minor_ticks = 5, visible = TRUE, number_formatter = c("basic",
  "numeral", "printf"), ...)
}
\arguments{
\item{fig}{figure to modify}

\item{label}{axis label}

\item{position}{where to place the axis (either "left" or "right")}

\item{log}{logical or integer - if TRUE, a log axis with base 10 is used - if an integer, a log axis with base of that integer will be used}

\item{grid}{logical - should a reference grid be shown for this axis?}

\item{num_minor_ticks}{number of minor ticks}

\item{visible}{should axis be shown?}

\item{number_formatter}{Bokeh numeric tick label formatter
(\href{http://bokeh.pydata.org/en/latest/docs/reference/models.html\#bokeh.models.formatters.BasicTickFormatter}{"basic"},
\href{http://bokeh.pydata.org/en/latest/docs/reference/models.html\#bokeh.models.formatters.NumeralTickFormatter}{"numeral"},
or \href{http://bokeh.pydata.org/en/latest/docs/reference/models.html\#bokeh.models.formatters.PrintfTickFormatter}{"printf"});
ignored if \code{log} is TRUE}

\item{\ldots}{additional parameters for fine control over axes (see "Additional parameters" below)}
}
\description{
Add y axis to a Bokeh figure
}
\section{Additional parameters}{

\subsection{Axis}{
\tabular{ll}{
\code{axis_label_standoff} \tab (integer) The distance in pixels that the axis labels should be offset from the tick labels. \cr
\code{axis_label_text_align} \tab ('left', 'right', 'center') The text align of the axis label. \cr
\code{axis_label_text_alpha} \tab (numeric) The text alpha of the axis label. \cr
\code{axis_label_text_baseline} \tab ('top', 'middle', 'bottom', 'alphabetic', 'hanging') The text baseline of the axis label. \cr
\code{axis_label_text_color} \tab (color) The text color of the axis label. \cr
\code{axis_label_text_font} \tab (string) The text font of the axis label. \cr
\code{axis_label_text_font_size} \tab (string - e.g. '12pt') The text font size of the axis label. \cr
\code{axis_label_text_font_style} \tab ('normal', 'italic', 'bold') The text font style of the axis label. \cr
\code{axis_line_alpha} \tab (numeric) The line alpha of the axis line. \cr
\code{axis_line_cap} \tab ('butt', 'round', 'square') The line cap of the axis line. \cr
\code{axis_line_color} \tab (color) The line color of the axis line. \cr
\code{axis_line_dash} \tab The line dash of the axis line. \cr
\code{axis_line_dash_offset} \tab (integer) The line dash offset of the axis line. \cr
\code{axis_line_join} \tab ('miter', 'round', 'bevel') The line join of the axis line. \cr
\code{axis_line_width} \tab (integer) The line width of the axis line. \cr
\code{major_label_orientation} \tab ('horizontal', 'vertical', or angle in degrees) What direction the major label text should be oriented. If a number is supplied, the angle of the text is measured from horizontal. \cr
\code{major_label_standoff} \tab (integer) The distance in pixels that the major tick labels should be offset from the associated ticks. \cr
\code{major_label_text_align} \tab ('left', 'right', 'center') The text align of the major tick labels. \cr
\code{major_label_text_alpha} \tab (numeric) The text alpha of the major tick labels. \cr
\code{major_label_text_baseline} \tab ('top', 'middle', 'bottom', 'alphabetic', 'hanging') The text baseline of the major tick labels. \cr
\code{major_label_text_color} \tab (color) The text color of the major tick labels. \cr
\code{major_label_text_font} \tab (string - 'Helvetica') The text font of the major tick labels. \cr
\code{major_label_text_font_size} \tab (string - e.g. '12pt') The text font size of the major tick labels. \cr
\code{major_label_text_font_style} \tab ('normal', 'italic', 'bold') The text font style of the major tick labels. \cr
\code{major_tick_in} \tab (integer) The distance in pixels that major ticks should extend into the main plot area. \cr
\code{major_tick_line_alpha} \tab (numeric) The line alpha of the major ticks. \cr
\code{major_tick_line_cap} \tab ('butt', 'round', 'square') The line cap of the major ticks. \cr
\code{major_tick_line_color} \tab (color) The line color of the major ticks. \cr
\code{major_tick_line_dash} \tab The line dash of the major ticks. \cr
\code{major_tick_line_dash_offset} \tab (integer) The line dash offset of the major ticks. \cr
\code{major_tick_line_join} \tab ('miter', 'round', 'bevel') The line join of the major ticks. \cr
\code{major_tick_line_width} \tab (integer) The line width of the major ticks. \cr
\code{major_tick_out} \tab (integer) The distance in pixels that major ticks should extend out of the main plot area. \cr
\code{minor_tick_in} \tab (integer) The distance in pixels that minor ticks should extend into the main plot area. \cr
\code{minor_tick_line_alpha} \tab (numeric) The line alpha of the minor ticks. \cr
\code{minor_tick_line_cap} \tab ('butt', 'round', 'square') The line cap of the minor ticks. \cr
\code{minor_tick_line_color} \tab (color) The line color of the minor ticks. \cr
\code{minor_tick_line_dash} \tab The line dash of the minor ticks. \cr
\code{minor_tick_line_dash_offset} \tab (integer) The line dash offset of the minor ticks. \cr
\code{minor_tick_line_join} \tab ('miter', 'round', 'bevel') The line join of the minor ticks. \cr
\code{minor_tick_line_width} \tab (integer) The line width of the minor ticks. \cr
\code{minor_tick_out} \tab (integer) The distance in pixels that major ticks should extend out of the main plot area. \cr
 }}
\subsection{Basic Tick Format}{
\tabular{ll}{
\code{power_limit_high} \tab (int) Limit the use of scientific notation to when log(x) >= value. \cr
\code{power_limit_low} \tab (int) Limit the use of scientific notation to when log(x) <= value. \cr
\code{precision} \tab (int) How many digits of precision to display in tick labels. Automatically
 determined if not specified. \cr
\code{use_scientific} \tab (logical) Whether to ever display scientific notation. If True, then when to use scientific
 notation is controlled by \code{power_limit_low} and \code{power_limit_high}. \cr
 }}
\subsection{Datetime Tick Format}{
\tabular{ll}{
\code{formats} \tab (list) Display tick values from a continuous range as formatted datetimes.
 See \href{http://bokeh.pydata.org/en/latest/docs/reference/models.html\#bokeh.models.formatters.DatetimeTickFormatter}{DatetimeTickFormatter}. \cr
 }}
\subsection{Numeral Tick Format}{
\tabular{ll}{
\code{format} \tab (string) Tick formatter based on a human-readable format string.
 See \href{http://bokeh.pydata.org/en/latest/docs/reference/models.html\#bokeh.models.formatters.NumeralTickFormatter}{NumeralTickFormatter}. \cr
 }}
\subsection{Printf Tick Format}{
\tabular{ll}{
\code{format} \tab (string) Tick formatter based on a printf-style format string.
 See \href{http://bokeh.pydata.org/en/latest/docs/reference/models.html\#bokeh.models.formatters.PrintfTickFormatter}{PrintfTickFormatter}. \cr
 }}
}
\examples{
\donttest{
figure() \%>\%
  ly_points(rexp(1000), rexp(1000)) \%>\%
  x_axis(label = "x", log = TRUE) \%>\%
  y_axis(label = "y", log = TRUE)

figure() \%>\%
  ly_points(2^(1:10)) \%>\%
  y_axis(log = 2)

# Disable scientific tick labels
figure() \%>\%
  ly_points(rnorm(10), rnorm(10) / 1000) \%>\%
  y_axis(use_scientific = FALSE)

# Specify datetime tick labels; the appropriate datetime units are automatically
#   chosen
figure() \%>\%
  ly_lines(seq(as.Date("2012-01-01"),as.Date("2012-12-31"), by="days"), 
           rnorm(366)) \%>\%
  x_axis(label = "Date", formats = list(months="\%b-\%Y", days="\%d"))
figure() \%>\%
  ly_lines(seq(as.Date("2012-01-01"),as.Date("2012-02-01"), by="days"), 
           rnorm(32)) \%>\%
  x_axis(label = "Date", formats = list(months="\%b-\%Y", days="\%d"))

# Specify numeric tick labels
figure() \%>\%
  ly_points(rnorm(10), rnorm(10) * 10000) \%>\%
  y_axis(number_formatter = "numeral", format = "0,000")

figure() \%>\%
  ly_points(rnorm(10), rnorm(10) * 100) \%>\%
  y_axis(number_formatter = "printf", format = "\%0.1f\%\%")
}
}
\seealso{
Other axes: \code{\link{x_axis}}
}

